AronGraphic

Documentation

data SegEndPt

Constructors

No

End

Beg

Both

Cen

All

data NormalDir

counter clockwise | clock wise

Constructors

NCCW

NCW

dist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute the distance between two points

\( v = (x, y, z) \)

\( \| v \| = \sqrt{ x^2 + y^2 + z^2} = \sqrt{ v \cdot v} \)

let v1 = Vertex3 1 2 3
let v2 = Vertex3 2 3 4 
dist v1 v2

sqdist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute the norm-squared

\( v = (x, y, z)\)

\( |v|^2 = x^2 + y^2 + z^2 \)

let v1 = Vertex3 1 2 3
let v2 = Vertex3 2 3 4 
sqdist v1 v2

dot3ve :: Num a => Vector3 a -> Vector3 a -> a

dot product for Vector3

dot3vx :: Num a => Vertex3 a -> Vertex3 a -> a

dot product for Vertex3

(+:) :: Num a => Vertex3 a -> Vector3 a -> Vertex3 a

(-:) :: Num a => Vertex3 a -> Vertex3 a -> Vector3 a

(⊥) :: Vector3 GLfloat -> Vector3 GLfloat

perpcw :: Num a => Vector3 a -> Vector3 a

perpccw :: Num a => Vector3 a -> Vector3 a

(*:) :: Num a => a -> Vector3 a -> Vector3 a

(**:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(*>:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(>>:) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

(.>) :: Int

(⊙) :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat

dot product odot

(∈) :: (Num a, Ord a) => a -> [a] -> Bool

(∘) :: Num a => a -> Vector3 a -> Vector3 a

(∑) :: Int

(∎) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

v2x :: Num a => Vector3 a -> Vertex3 a

x2v :: Num a => Vertex3 a -> Vector3 a

rf :: (Real a, Fractional b) => a -> b

Shorten realToFrac

red :: Color3 GLdouble

------------------------------------------------------------------------------- | Fri Dec 7 14:35:38 2018 | three colors: data Color3 a = Color3 !a !a !a

green :: Color3 GLdouble

blue :: Color3 GLdouble

skew :: Num a => Vertex3 a -> [[a]]

form a skew symmetric matrix from Vertex3 | | cross product of a x b = [a] b where [a] is skew symmetric matrix from vector a | colIndex x rowIndex = [c][r] | i j k k i | i -1 -k j j | j k -1 -i | k -j i -1

skew' :: Num a => Vector3 a -> [[a]]

(⊗) :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

Cross product of two vectors.

The directin is determinated by the the Right Hadle Rule

isCoplanar :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

check whether four points are coplanar

| Userscatmyfilebitbucketmathcoplanar.pdf

cross :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

normal3 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vector3 GLfloat)

compute the normal of three points

normal3' :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

compute the normal of plane with two given vectors

vec :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat

Vector: p0 -> p1 = p1 - p0

ray :: Vertex3 GLfloat -> GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat

Given a point and vector,

| draw a line passes the point and alone the vector

| r(t) = p0 + t(p1 - p0)

perpenLine :: GLfloat -> GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat]

Given s, t, p0 and p1 1. c₀ = center p₀ p₁ 2. p1' = p0 + s(⊥ p0 p1), p2' = p0 + t(⊥ p0 p1) 3. draw line from p1' to p2'

curvePt :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> [Vertex3 GLfloat]

drawCurve :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> Color3 GLdouble -> IO ()

Given a function f, interval (a, b), draw the curve from a to b

tangentLine :: Fractional a => (a -> a) -> (a, a) -> a -> [Vertex3 a]

Draw a line that is tangent at pt (c, f(c)) where x in [x0, x1]

tangentLine

Given a function f, and x0 on x-Axis

Derive a tangent line at (c, f(c)) with slop = f'(c)

Interpolate (x0, f(x0)) and (x1, f(x1))

interpolate' :: Fractional a => Integer -> Vertex3 a -> Vertex3 a -> [Vertex3 a]

drawTangentLine :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> Color3 GLdouble -> IO ()

Given a function f, interval (a, b), draw a tangent line at (x0, f(x0)) from a to b f::x -> y (a, b) is interval differentiate at (x0, f x0)

normalLine :: NormalDir -> (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> [Vertex3 GLfloat]

NormalDir is either counter clockwise or clockwise for normal | (x0, x1) is interval for the normal | f is any function | x=c is the tangent point at (c, f c) |

drawNormalLine :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> Color3 GLdouble -> IO ()

cen :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

compute the center of two vertices/vertex

isPerpen :: Vector3 GLfloat -> Vector3 GLfloat -> Bool

check two vectors whether they are perpendicular

isColinear :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

Tue Dec 11 16:05:12 2018 | check three pts whether they are colinear or not | | f(t) = p0 + t(p1 - p0) | f(s) = p0 + s(p2 - p0) | f(t) = f(s) => | t(p1 - p0) = s(p2 - p0) | t v1 = s v2, let A = [v1 v2] | det A =? 0 | det A = 0 => p0 p1 p2 are colinear |

isInSegment :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe Bool

given p0, q0 q1, check whether p0 is inside the segment of q0 q1

drawPrimitive :: PrimitiveMode -> Color3 GLdouble -> [(GLfloat, GLfloat, GLfloat)] -> IO ()

draw primitive such lines, points, lineloop etc

drawPrimitive' :: PrimitiveMode -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawLines :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegmentWithEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegmentNoEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegments :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

Draw a set of segments | Segment contains two end points, one is begin point, other is end point [p0, p1] = p0 -> p1

drawSegment :: Color3 GLdouble -> (Vertex3 GLfloat, Vertex3 GLfloat) -> IO ()

Draw one segment from p0 to p1 | drawSegment color (p0, p1) |

drawSegment' :: SegEndPt -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

definition: Segment contains two end points, one is begin point, other is end point [p0, p1] = p0 -> p1

data SegEndPt = No -- no pt, just a segment | End -- end pt | Beg -- begin pt | Both -- begin and end pts | Cen -- center pt | All -- all pts: begin, end and ceneter

drawSegmentArg :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

Draw intersection pt from two segments | Two segments need not have an intersection.

drawSpherePt :: Vertex3 GLfloat -> GLfloat -> [Vertex3 GLfloat]

draw sphere center at c with radius r

β is in x-y plane, it rotates around z-Axis. ϕ is x-z plane, it rotates aroound y-Axis

\( r*\cos(ϕ) \) is the radius of circle cut through x-y plane

drawPlane :: Color3 GLdouble -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

draw plane with three points | no check whether they are co-linear

circle' :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

set of pts for circle circle::(GLfloat, GLfloat, GLfloat) -> Double -> [(GLfloat, GLfloat, GLfloat)] circle (x0, y0, z0) r =[ let alpha = (pi2*n)/num in ((rf r)*_sin(alpha) + x0, (rf r)*_cos(alpha) + y0, 0 + z0) | n <- [1..num]] where num = 40 pi2 = 2*pi::Float

draw xy-plane circle | divide 2*pi to num and iterate from 1 to num |

drawDot :: Vertex3 GLfloat -> IO ()

draw dot, small circle

cartToPolar :: Vector3 GLfloat -> (GLfloat, GLfloat)

Given an Vector3 x y z or Vertex3 x y z | Convert Cartesian Coordinates to Polar Coordinates | beta x-y plane, | alpha x-z plane | | y z | | / | ---|/--- x | /

roty :: Floating a => a -> [[a]]

right hand coordinates system in OpenGL | | x-z plane

rotz :: Floating a => a -> [[a]]

x-y plane

drawCircleVec :: GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat -> IO ()

Given an radius r, center Vertex3 vx, and Vector ve Draw circle(r, vx) which is perpendicular to ve |

matVx :: [[a]] -> Vertex3 a

matrix to Vertex

matVe :: [[a]] -> Vector3 a

matrix to Vector

veMat :: Vector3 a -> [[a]]

Vector to matrix

drawCircle' :: Vertex3 GLfloat -> Double -> IO ()

draw circle with center and radius | ceneter radius drawCircle::(GLfloat, GLfloat, GLfloat) -> Double -> IO() drawCircle cen r = drawPrimitive LineLoop red $ circle cen r

drawCircleColor :: Vertex3 GLfloat -> Color3 GLdouble -> Double -> IO ()

drawCircleColor' :: Color3 GLdouble -> Double -> Vertex3 GLfloat -> IO ()

conic :: [(GLfloat, GLfloat, GLfloat)]

Conic Parameter Equation | | gx http://localhost/html/indexThebeautyofTorus.html |

det2 :: Num a => [[a]] -> a

determinant of two dimension matrix

inverseR :: [[Rational]] -> [[Rational]]

inv2 :: Fractional a => [[a]] -> [[a]]

inverse of two dismension matrix [a b] inv 1/det [d -b] [c d] [-c a]

intersectLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 GLfloat, [[GLfloat]])

data Seg a = Seg a a | fun::Seg GLflat -> Seg GLfloat -> Maybe(Vertex3 GLfloat, [[GLfloat]]) | fun (Seg (Vertex3 x0 y0 z0) (Vertex3 x1 y1 z1)) = Nothing

find the intersection of two line | if p0 p1 and q0 q1 are not parallel, | return the vertex of intersection and s and t | gx http://localhost/pdf/intersectionLine.pdf

let inter = let p0 = Vertex3 (x0=0) (y0=0) (z0=0) p1 = Vertex3 (x1=1) (y1=1) (z1=0) q0 = Vertex3 (a0=2) (b0=0) (c0=0) q1 = Vertex3 (a1=2) (b1=1) (c1=0) in intersectLine p0 p1 q0 q1

intersectLine' :: [Vertex3 GLfloat] -> Maybe (Vertex3 GLfloat, [[GLfloat]])

Four Vertex3 in a list

pointToLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

TODO: fix, the code only handles two dimension. | Given a point: p0, line: q0 q1 | Compute the distance from p0 to line: q0 q1 |

intersectedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat)

find the intersection of two segments, otherwise return Nothing

isIntersectedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Bool

intersectLineR :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational, [[Rational]])

threePtDeterminant :: Fractional a => Vertex3 a -> Vertex3 a -> Vertex3 a -> a

Three points rotation order

If given three points in following order:

p₀ = (1, 0)

p₁ = (0, 0)

p₂ = (1, 0)

then vectors are computed in following:

v10 = p₀ - p₁

v12 = p₂ - p₁

matrix can be formed as following:

m = [v10, v12]

\( \vec{v_{01}} \times \vec{v_{02}} \) in Right Hand Rule

\[ m = \begin{bmatrix} 1 & 0 \\ 0 & 1 \\ \end{bmatrix} \\ \det m = 1 > 0 \\ \]

If the three points are collinear:

\( \det M = 0 \)

If the order of three points in clockwise order:

\( \det M > 0 \)

If the order of three points in counter clockwise order:

\( \det M < 0 \)

threePtCCW :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points in Counter ClockWise order

\( \det M > 0 \)

threePtCW :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points in ClockWise order

\( \det M < 0 \)

threePtCollinear :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points are collinear

\( \det M = 0 \)

ptInsideTri :: Vertex3 GLfloat -> (Vertex3 GLfloat, Vertex3 GLfloat, Vertex3 GLfloat) -> (Bool, GLfloat)

Check whether a point p₀ is inside a triangle \( \triangle ABC \)

NOTE the order of three pts: a, b, c does't matter. e.g CCW or CW

p₀ can NOT be the same point as a, b or c

If p₀ is collinear with AB, BC, or AC, then p₀ is inside the triangle

p0 is the point that is tested three points a b c forms a triangle

is return true if point p0 inside the triangle

The sum of three angles are in degree.

let p0 = Vertex3 0.1 0.1 0
let a  = Vertex3 1 0 0
let b  = Vertex3 0 0 0
let c  = Vertex3 0 1 0

ptInsideTri p0 a b c
(True, 360.0)

perpPlane :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Given a point p0 and three pts: q0, q1, q2,

Compute the intersection of line is perpendicular to the plane and passes point p0

If three pts (q0, q1, q2) are colinear, return Nothing

otherwise, return Just (Vertex3 GLfloat)

lineIntersectPlane :: Vertex3 GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Line intersects Plane | Given Line: p0 p1 and Plane: q0 q1 q2 | Compute the intersection of line and plane | p1 = p0 +: h(p1 -: p0)

convexHull :: Int -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat] -> [[Vertex3 GLfloat]]

Naive algorithm to compute the convex hull

convexHull n top' top pts
n the total of vertices
top' = Vertex3 x+1 y z
top  = Vertex3 x y z
pts is the list of vertices

ConvexHull

algorithm

find a top vertex: B(Vertex3 x y z) along y-Axis

construct a vertex: A(Vertex3 x+1 y z) according to top(Vertex3 x y z)

compute the angle \( \cos ∠ABC \) from three points: A B C where C is one of vertices list excluding B

Note: Assume the number of pts is three or more

If there are two pts for the list of Vertex3 e.g. p0 p1

then draw line(p0 p1) and line(p1 p0) so there is duplicated line

Initially, xi = top

n = length pts => n vertices => n edges. if there is two vertices, then (p0, p1) and (p1, p0)

xi is dumy init value, it is the top pt alone y-Axis

top' is the top=Vertex3 x y z => top'=(Vertex3 (x+1) y z)

top is the top pt for alone y-Axis

vec(top x1) and vec(top x2)

v1 = top -> x1 = x1 -: top

v2 = top -> x2 = x2 -: top

TODO: Handle the case where the three points are colinear.

projv :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

Projection from \(u\) onto \(v\) in Vector3

projection

nr :: Vector3 GLfloat -> GLfloat

Compute the norm of a vector

\(v = (x, y, z)\)

\(|v| = \sqrt{ x^2 + y^2 + z^2} \)

uv :: Vector3 GLfloat -> Vector3 GLfloat

Normalize a vector

e.g \( \|\vec{v}\| = 1 \)

rod :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat -> Vector3 GLfloat

Rodrigue formula

\(u\) rotates around \(v\) in angle \(\phi\) in right hand rule

rejection = \(u - p\)

gx Rotation

cosVex3 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute an angle from three points: \(a, b, c \) with dot product

let a = Vertex3 1 0 0
let b = Vertex3 0 0 0
let c = Vertex3 1 0 0
cosVex3 a b c
0.0

\( \vec{ba} = a - b \)

\( \vec{bc} = c - b \)

\[ \begin{equation} \begin{aligned} \vec{ba} \circ \vec{bc} &= | \vec{ba} | | \vec{bc} | \cos{\beta} \\ \cos{\beta} &= \frac{ \vec{bc} \circ \vec{bc} }{| \vec{ba} | | \vec{bc}|} \\ \end{aligned} \end{equation} \]

Orphan instances

Num a => Num (Vector3 a)
(Vector3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vector3 a -> Vector3 a -> Vector3 a

(-) :: Vector3 a -> Vector3 a -> Vector3 a

(*) :: Vector3 a -> Vector3 a -> Vector3 a

negate :: Vector3 a -> Vector3 a

abs :: Vector3 a -> Vector3 a

signum :: Vector3 a -> Vector3 a

fromInteger :: Integer -> Vector3 a

Num a => Num (Vertex3 a)
(Vertex3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(-) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(*) :: Vertex3 a -> Vertex3 a -> Vertex3 a

negate :: Vertex3 a -> Vertex3 a

abs :: Vertex3 a -> Vertex3 a

signum :: Vertex3 a -> Vertex3 a

fromInteger :: Integer -> Vertex3 a