AronModule

Documentation

class Cnum a where

load shell environments inside GHC | use the SAME NAME as in shell environments

Overloading the sin, cos to avoid Haskell crazy strict type

Minimal complete definition

_sqrt, _sin, _cos

Methods

_sqrt :: a -> Float

_sin :: a -> Float

_cos :: a -> Float

class Dnum a b where

Division is painful in Haskell

| Try to overload _div for many types

Minimal complete definition

_div

Associated Types

type DivType a b

Methods

_div :: a -> b -> DivType a b

class Add a b where

Minimal complete definition

addd

Associated Types

type SumTy a b

Methods

addd :: a -> b -> SumTy a b

div' :: Int -> Int -> Double

mat :: [[Integer]]

simple matrix can be used in GHCi

mat' :: [[[Char]]]

data Quaternion

Quaternion Implementation

Constructors

Quaternion

Fields

a :: Float

b :: Float

c :: Float

d :: Float

data Complex

Implement complex number operations

Constructors

Complex

Fields

x :: Float

y :: Float

conjugate :: Complex -> Complex

magnitude :: Complex -> Float

realPart :: Complex -> Float

imagPart :: Complex -> Float

coord :: Complex -> (Float, Float)

radianToDegree :: Float -> Float

degreeToRadian :: Float -> Float

tri :: Complex -> (Float, Float, Float)

data GPoint

Constructors

GPoint Int Int

getX :: GPoint -> Int

getY :: GPoint -> Int

cat :: [String] -> String

replaceList :: [String] -> String -> String -> [String]

["cat", "dog"] "cat" "pig" => ["pig", "dog"]

| replace a matching string with other string

repeat' :: Integer -> a -> [a]

repeat's 3 "pig" => ["pig", "pig", "pig"]

replicate' :: Int -> a -> [a]

replicate' 4 a => "aaaa"

catMaybe :: Maybe [Int] -> Maybe [Int] -> Maybe [Int]

removeSpace :: String -> String

isEmpty :: String -> Bool

trimWS :: String -> String

trim :: String -> String

splitListEmptyLine :: [String] -> [[String]]

split list of string when string is empty/space | ["dog", " ", "", "cat"] => [["dog"], [], ["cat"]]

splitList :: [String] -> ([String], [String])

pa :: Show a => [[a]] -> IO ()

pa1 :: Show a => [a] -> IO ()

partList :: Int -> [a] -> [[a]]

partList2 :: Int -> [a] -> [[a]]

partitionStrRegex :: String -> String -> [String]

partition string to [String] according to character class [] | r = "[,.]" s = "dog,cat,cow.fox" => ["dog", "cat", "cow", "fox"]

prefix :: String -> [String]

unique :: Ord a => [a] -> [a]

mergeListList :: [[String]] -> [[String]] -> [[String]]

mergeList :: [a] -> [a] -> [a]

mergeListLen :: [a] -> [a] -> Maybe [a]

iterateList :: [a] -> (a -> IO ()) -> IO ()

unwrap :: Maybe a -> a

codeCapture :: String -> String

qqsort :: (a -> a -> Bool) -> [a] -> [a]

There are many versions of quicksort here

bad version

quickSort :: [Int] -> [Int]

better version

quickSort' :: [Int] -> [Int]

nice version

quickSort1 :: Ord a => [a] -> [a]

mergeSort :: [Int] -> [Int]

mergeSortM :: [[Integer]] -> [[Integer]]

readFileToList :: FilePath -> IO [String]

writeToFile :: FilePath -> [String] -> IO ()

wfs :: FilePath -> [String] -> IO ()

wf :: FilePath -> String -> IO ()

writeToFileAppend :: FilePath -> [String] -> IO ()

pp :: Show s => s -> IO ()

fl :: IO ()

ff :: Show var => String -> var -> IO ()

fw :: String -> IO ()

listDot :: Num a => [a] -> [a] -> a

scaleList :: Num a => a -> [a] -> [a]

listScale :: Num a => [a] -> a -> [a]

listAdd :: Num a => [a] -> [a] -> [a]

listSub :: Num a => [a] -> [a] -> [a]

listMul :: Num a => [a] -> [a] -> [a]

listNeg :: Num a => [a] -> [a]

dim :: [[a]] -> (Int, Int)

get dimenstion of matrix -> (nrow, ncol) | if [] -> (0, 0)

removeIndex :: Int -> [a] -> [a]

removeRowCol :: Integer -> Integer -> [[Integer]] -> [[Integer]]

multiMatInt :: [[Int]] -> [[Int]] -> [[Int]]

matrix multiplicaiton in Int

multiMatDouble :: [[Double]] -> [[Double]] -> [[Double]]

matrix multiplication in Double

multiMat :: Num a => [[a]] -> [[a]] -> [[a]]

matrix multiplication in any type Num

multiMatVec :: Num a => [[a]] -> [a] -> [a]

multiRatMat :: [[String]] -> [[String]] -> [[String]]

matrix multiplication in String/Rational

sumRatList :: [String] -> String

addR :: String -> String -> String

red1 :: String -> String

red2 :: String -> String

reduceForm :: [[String]] -> [[String]]

reduce :: [Integer] -> [Integer]

stringToInt :: String -> Integer

intToString :: Integer -> String

splitR :: String -> [Integer]

multR :: String -> String -> String

multRL :: String -> [String] -> [String]

normR :: [String] -> [String] -> String

proj :: [String] -> [String] -> [String]

addRL :: [String] -> [String] -> [String]

divR :: String -> String -> String

divR' :: Integer -> Integer -> String

invR :: String -> String

subR :: String -> String -> String

negR :: String -> String

negList :: [String] -> [String]

subRL :: [String] -> [String] -> [String]

diag1 :: Num a => [[a]] -> [a]

get the diagonal of the matrix

uptri :: Num a => [[a]] -> [[a]]

get upper triangle of matrix

trans :: [[a]] -> [[a]]

transpose the matrix

data Vec a

Constructors

Vec [String]

data Rat a

Constructors

Rat String

(⨂) :: Vec a1 -> Vec a2 -> Rat a3

unicode operator | dot product of two vectors

(⨁) :: Vec a1 -> Vec a2 -> Vec a3

addition of two vectors

(⦵) :: Vec a1 -> Vec a2 -> Vec a3

subtraction of two vectors

(⨸) :: Rat a1 -> Rat a2 -> Rat a3

division of two rational numbers

sortRow :: [[Integer]] -> [[Integer]]

sort row for list of list

upperTri :: [[Integer]] -> [[Integer]]

tranform the matrix to echolon/upper triangle form | Note: the determinant is different than the original matrix

divI :: Integer -> Integer -> Double

inverse :: [[Integer]] -> ([[Double]], [[String]])

Find the invertible matrix, return ([[]], [[]]) if the matrix is singular | The code does not check whether the matrix is singular or not | I'm sure there is better way to it better

isInvertible :: [[Integer]] -> Bool

Multiply all the diagonal elements and check whether the product is zero or not

ident :: Integer -> [[Integer]]

Generate identity matrix

ident' :: Int -> [[Integer]]

identS :: Integer -> [[String]]

help :: IO ()

cmd :: [String] -> IO ()

compileHaskellToBin :: String -> String -> IO ()

Compile haskell code to $ffmybinmyHaskell => create symbol link $sym/sym | [file:myHaskell.hs] [sym:symbol] link name in $sym

drawInt :: Int -> Int -> IO Int

Gene random number

randomList :: Int -> IO [Integer]

randomMatrix :: Int -> Int -> IO [[Integer]]

generate m x n random matrix

ls :: IO ()

Try to replace as many as shell command as possible | See how far I can go | write many shell commands as possible try to emulate shell commands

lsFile :: String -> IO [String]

lsFileFull :: String -> IO [String]

lsRegex :: String -> String -> IO [String]

lsFullRegex :: String -> String -> IO [String]

rm :: FilePath -> IO ()

rmDir :: FilePath -> IO ()

pwd :: IO ()

getPwd :: IO FilePath

cd :: FilePath -> IO ()

en :: String -> IO String

cc :: String -> IO ()

g :: IO ()

asplitPath :: FilePath -> [String]

pathBase :: FilePath -> FilePath

copyFileToDir :: FilePath -> FilePath -> IO ()

isFile :: FilePath -> IO Bool

fExist :: FilePath -> IO Bool

fe :: FilePath -> IO Bool

copyDir :: FilePath -> FilePath -> IO ()

copy dir to other dir

mv :: FilePath -> FilePath -> IO ()

renameAllFile :: String -> String -> IO ()

mvFiles :: String -> String -> IO ()

mvFile :: FilePath -> FilePath -> IO ()

mkdir :: FilePath -> IO ()

mkdirp :: FilePath -> IO ()

sfilter :: String -> [String] -> [String]

run :: String -> IO [String]

df :: Fractional a => (a -> a) -> a -> a

Some Math function | derivative of f | f'(x) = lim_{h rightarrow 0} f(x + h) - f(x)/h

tangentLine :: Fractional a => (a -> a) -> a -> a -> a

draw tangent line at point (x0, f(x0)) for function f

prime :: [Integer]

generate prime number with Sieve Algorithm

oneRoot :: (Double -> Double) -> Double -> Double -> Double -> Maybe Double

------------------------------------------------------------------------------- Find root for any function

Partition the interval [x0, xn] into list = [x0, x1) [x1, x2) [x2, x3) ..[x(n-1), xn) Cocat [xn, xn] with the list since the left boundary is checked only | Note: f(xn) might be zero, we need to check the xn boundary | TODO: need to check f(xn), DONE, concat [xn, xn] to the list | -------------------------------------------------------------------------------- | Good test cases: | f(x) = x^2 - 4, [-2, 2] | f(x) = x^5 -4*x^4 + 0.1*x^3 + 4*x^2 - 0.5, [-4, 4] => 5 solutions | -------------------------------------------------------------------------------- | limitation: | if each subinterval contains two or more values, then ONLY one value can be found | subinterval can be adjusted in arbitrary small | -------------------------------------------------------------------------------- | 0, 2/2 = [0, 1, 2] = [0, 1)[1, 2)

rootList :: (Double -> Double) -> Double -> Double -> Double -> Integer -> [Maybe Double]

Find all the roots for a given close interval: [1, 2], 1 or 2 might be the root |