WaiLib

Description

The Module contains all the functios for wai.hs

Use Aeson to serialize record to Json

Record: Person

Insert data to MySqlit-simple file-based database

Upload file to server.

Use Redis(memcached) to store snippet and query snippet.

Documentation

data Block

Constructors

Block

Fields

bblock :: [Text]

data MBlock

Constructors

MBlock

Fields

mblock :: [Integer]

data GeneMatrix

Constructors

GeneMatrix

Fields

cmd :: Text

ncol :: Integer

nrow :: Integer

data ReplyCode

Constructors

ReplyCode

Fields

rcmd :: Text

rerror :: Text

stdout :: Text

data CompileCode

Constructors

CompileCode

Fields

compiler :: Text

option :: Text

code :: Text

data MatInt

Constructors

MatInt

Fields

name :: Text

matrix :: [[Integer]]

data Person

Person to Json object

Constructors

Person

Fields

personId :: Int64

personName :: Text

personAge :: Text

data UserInput

create UserInput table in Sqlite

cmdId = pid

xcmd = input command, e.g. "c ls"

Constructors

UserInput

Fields

cmdId :: Int64

xcmd :: Text

updir :: String

runSh :: Text -> IO (ExitCode, Text, Text)

styleChar :: String -> String -> Char -> Char -> String -> String

href :: String -> String -> String -> String

------------------------------------------------------------------------------- | Thu Nov 15 13:18:04 2018 | Simple web server for request and response | Handle search code snippet | -------------------------------------------------------------------------------- run.sh => ./wai ghc -i$bhaskelllib --make wai.hs -o wai [file path] [desc] [image src]

changeSymbol :: String -> String

listToByteStr :: [[String]] -> ByteString

Convert [[String]] to ByteString

cssStrong :: String -> String

cssHead :: [[String]] -> [[String]]

htmlLess :: [[String]] -> [[String]]

htmlGreater :: [[String]] -> [[String]]

keyWord :: [[String]] -> [[String]]

keyWord1 :: [[String]] -> [[String]]

keyDash :: [[String]] -> [[String]]

keySymbol1 :: [[String]] -> [[String]]

keyName :: [[String]] -> [[String]]

specialName :: [[String]] -> [[String]]

javaClassName :: [[String]] -> [[String]]

javaFunClass :: [[String]] -> [[String]]

javaKeyWords :: [[String]] -> [[String]]

javaCmdKeyWords :: [[String]] -> [[String]]

mysqlKeyWords :: [[String]] -> [[String]]

keyURL :: [[String]] -> [[String]]

spChar :: [[String]] -> [[String]]

bracketChar :: [[String]] -> [[String]]

sbChar :: [[String]] -> [[String]]

transform :: [[String]] -> [[String]]

hiddenForm :: Integer -> String -> String

Hide all the data in TextArea

   action="sendConfirmation.php" name="confirmationForm" method="post"
   id="confirmationText" class="text" cols="86" rows ="20" name="confirmationText"/textarea

   type="submit" value="Email" class="submitButton"
   /form

   cols="20" rows="20" id="textArea" style="display:none;font-size:18px;" class="hide"/textarea

   autofocus="true" onfocus="textAreaAdjust(this);"/textarea
   

foldListList :: ([[String]] -> [[String]]) -> [[String]] -> String

foldListListTxt :: [[String]] -> String

preT :: [Char] -> [Char]

divT :: [Char] -> [Char]

divC :: [Char] -> [Char]

tableT :: [Char] -> [Char]

tableC :: [Char] -> [Char]

tableAC :: [Char] -> [Char] -> [Char]

tdT :: [Char] -> [Char]

tdC :: [Char] -> [Char]

tdAC :: [Char] -> [Char] -> [Char]

trC :: [Char] -> [Char]

id_ :: [Char] -> [Char]

class_ :: [Char] -> [Char]

onclick_ :: [Char] -> [Char]

style_ :: Foldable t => t [Char] -> [Char]

align_ :: [Char] -> [Char]

table :: [[String]] -> String

pdfname :: String

img :: String

wapp :: FilePath

pdfPath :: [Char]

docRoot :: [Char]

doc :: FilePath

cmdLog :: FilePath

currCmdFile :: FilePath

logCurrCmd :: [String] -> IO ()

readCurrCmd :: IO String

type HMap = HashMap String [[String]]

genePDF :: String -> IO ()

app :: Connection -> Connection -> IORef HMap -> Application

plainIndex :: Response

pdfFile :: Response

insertinfo :: Response

searchUI :: Response

notFound :: Response

notFoundStr :: ByteString -> Response

snippetPath :: FilePath

insertAll :: [(String, [[String]])] -> HMap -> HMap

mapClear :: [String] -> HMap -> HMap

snippetMap :: [([String], [String])] -> IORef HMap -> IO ()

read snippet file

NOTE The code can be speed up a bit, change [String] [[String]

type HMap = M.HashMap String [[String]] => type HMap = M.HashMap String (Set [String])

spanBlock :: HMap -> Maybe ByteString -> String

http://localhost:8080/snippet?id=keyinput | Conver ByteString to String or vice versa | span block code: spantext/span | | type in search key | transform a block of code => colourful block of code | each block of code => key -> "block of code" | return a block of colourful code

spanBlockX :: ([[String]] -> [[String]]) -> HMap -> Maybe ByteString -> String

spanBlockXX :: HMap -> Maybe ByteString -> String

(∘) :: [a] -> [a] -> [a]

htmlForm :: String -> String

htmlPre :: String -> String

htmlBody :: String -> String

replyHtml :: String -> String -> String

anyRoute :: Connection -> IORef HMap -> Request -> Response

snippet?id=queryStr S8.unpack: ByteString to String type Application = Request -> (Response -> IO ResponseReceived) -> IO ResponseReceived anyRoute => Response

uploadPage :: Response

http://localhost:8080/up/ | NOTE: file => /upload dir | Plz see uploadPage.html | Userscatmyfilebitbuckethaskell_webappuploadPage.html | responseFile :: H.Status -> H.ResponseHeaders -> FilePath -> Maybe FilePart -> Response

data SearchType

Constructors

CmdT

JavaT

HaskellT

SnippetT

readCmd :: FilePath -> SearchType -> IO String

groupCountFilter :: [String] -> [String]

Filter some commands out from a list.

optionHtml :: [String] -> String

responseNothing :: String -> Response

responseNothingBS :: ByteString -> Response

responseCmd :: Connection -> String -> Response

Response output from shell command

The maximum number of lines are 200, " | head -200"

Exit code can not be checked, ExitSuccess

Wed May 8 23:10:41 2019 3 Error can be checked in stderr >(e, so, si) <- A.runSh $ toText (drop 2 ncmd) > if si is NOT empty then there is stderr

type streamBody = (Builder -> IO()) -> IO() -> IO()

responseXCmd :: String -> Response

responseJavaHtml :: String -> Response

responseHaskellHtml :: String -> Response

queryLibHaskell :: String -> Response

query function info from redis without Html

query_redis = "/Users/cat/myfile/symbin/redis_query.hs "

preKey = "AronModule."  AronModule.hs
preKey = "Aron."        Aron.java

queryLibJava :: String -> Response

responseSnippet :: Connection -> String -> IORef HMap -> Response

responseSnippetTxt :: String -> IORef HMap -> Response

query snippet from HMap without Html

trimBothLBS :: ByteString -> ByteString

geneRectMat :: Application

receiveCode :: Application

receiveCode2 :: Application

responseEditor :: Response

responseJavascript :: FilePath -> Response

responseHtml :: FilePath -> Response

updateMap :: IORef HMap -> Application

   action="/upload" method="POST" enctype="multipart/form-data"
   Upload File: type="file" name="file"br
   type="submit" value="submit"
   /form 
   

Handle file uploads, storing the file in the current directory

upload :: Application

replyEditor :: Response

NOTE: file => /upload dir | Plz see uploadPage.html

responseHelp :: Response

replyCssButton :: Response

wordcountReply :: Response

matrixReply :: Response

replyJS :: Response

insertDatabase :: Connection -> Application

Insert name and age to MySqlite-simple file-based database.

http://localhost:8080/insert/

File: insert.html action="/insert" method="POST" enctype="multipart/form-data" Name type="text" name="name"br Age type="text" name="age"br type="submit" value="submit" /form

insert data to table: people
"INSERT INTO people (name, age) VALUES (?,?)" 

upload :: String -> Application

------------------------------------------------------------------------------- | Wed Dec 5 15:06:00 2018 Sat Jun 8 23:42:18 2019 | upload with following POST to upload file to server | -------------------------------------------------------------------------------- action="/upload" method="POST" enctype="multipart/form-data" Upload File: type="file" name="file"br type="submit" value="submit" /form | -------------------------------------------------------------------------------- http://localhost:8080/up/ | File is uploaded to => haskell_web/uploaddir

searchMap :: Application

type ClientId = Int

type Client = (ClientId, Connection)

type State = [Client]

connectClient :: Connection -> MVar State -> IO ClientId

withoutClient :: ClientId -> State -> State

disconnectClient :: ClientId -> MVar State -> IO ()

nextId :: State -> ClientId

httpApp :: Application

listen :: Connection -> ClientId -> MVar State -> IO ()

broadcast :: ClientId -> MVar State -> Text -> IO ()

wsApp :: MVar State -> ServerApp