AronGraphic

Documentation

data SegEndPt

Constructors

No

End

Beg

Both

Cen

All

rf :: (Real a, Fractional b) => a -> b

Shorten realToFrac

red :: Color3 GLdouble

------------------------------------------------------------------------------- | Fri Dec 7 14:35:38 2018 | three colors: data Color3 a = Color3 !a !a !a

green :: Color3 GLdouble

blue :: Color3 GLdouble

dot3vx :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

dot product for Vertex3 GLfloat

skew :: Vertex3 GLfloat -> [[GLfloat]]

form a skew symmetric matrix from Vertex3 | | cross product of a x b = [a] b where [a] is skew symmetric matrix from vector a | colIndex x rowIndex = [c][r] | i j k k i | i -1 -k j j | j k -1 -i | k -j i -1

skew' :: Vector3 GLfloat -> [[GLfloat]]

(⊗) :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

cross product in right hadle rule

isCoplanar :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

check whether four points are coplanar

| Userscatmyfilebitbucketmathcoplanar.pdf

cross :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

normal3 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vector3 GLfloat)

compute the normal of three points

normal3' :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

compute the normal of plane with two given vectors

dist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

sqdist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

dot3ve :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat

(+:) :: Vertex3 GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat

(-:) :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat

(⊥) :: Vector3 GLfloat -> Vector3 GLfloat

perv :: Num a => Vector3 a -> Vector3 a

perx :: Num a => Vertex3 a -> Vertex3 a

(*:) :: Num a => a -> Vector3 a -> Vector3 a

(**:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(*>:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(>>:) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

(.>) :: Int

(⊙) :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat

dot product odot

(∈) :: (Num a, Ord a) => a -> [a] -> Bool

(∘) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

(∑) :: Int

(∎) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

v2x :: Vector3 GLfloat -> Vertex3 GLfloat

x2v :: Vertex3 GLfloat -> Vector3 GLfloat

smulv :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

vec :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat

Vector: p0 -> p1 = p1 - p0

perpenLine :: GLfloat -> GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat]

Given s, t, p0 and p1 1. c₀ = center p₀ p₁ 2. p1' = p0 + s(⊥ p0 p1), p2' = p0 + t(⊥ p0 p1) 3. draw line from p1' to p2'

cen :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

compute the center of two vertices/vertex |

isPerpen :: Vector3 GLfloat -> Vector3 GLfloat -> Bool

check two vectors whether they are perpendicular

isColinear :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

Tue Dec 11 16:05:12 2018 | check three pts whether they are colinear or not | | f(t) = p0 + t(p1 - p0) | f(s) = p0 + s(p2 - p0) | f(t) = f(s) => | t(p1 - p0) = s(p2 - p0) | t v1 = s v2, let A = [v1 v2] | det A =? 0 | det A = 0 => p0 p1 p2 are colinear |

isInSegment :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe Bool

given p0, q0 q1, check whether p0 is inside the segment of q0 q1

drawPrimitive :: PrimitiveMode -> Color3 GLdouble -> [(GLfloat, GLfloat, GLfloat)] -> IO ()

draw primitive such lines, points, lineloop etc

drawPrimitive' :: PrimitiveMode -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawLine :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegmentWithEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegmentNoEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegment :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

definition: Segment contains two end points, one is begin point, other is end point [p0, p1] = p0 -> p1

drawSegment' :: SegEndPt -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

definition: Segment contains two end points, one is begin point, other is end point [p0, p1] = p0 -> p1

drawSegmentArg :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

Draw intersection pt from two segments | Two segments need not have an intersection.

drawSpherePt :: Vertex3 GLfloat -> GLfloat -> [Vertex3 GLfloat]

draw sphere center at c with radius r | β is in x-y plane, ϕ is

drawPlane :: Color3 GLdouble -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

draw plane with three points | no check whether they are co-linear

circle' :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

set of pts for circle circle::(GLfloat, GLfloat, GLfloat) -> Double -> [(GLfloat, GLfloat, GLfloat)] circle (x0, y0, z0) r =[ let alpha = (pi2*n)/num in ((rf r)*_sin(alpha) + x0, (rf r)*_cos(alpha) + y0, 0 + z0) | n <- [1..num]] where num = 40 pi2 = 2*pi::Float

draw xy-plane circle | divide 2*pi to num and iterate from 1 to num |

drawDot :: Vertex3 GLfloat -> IO ()

draw dot, small circle

cartToPolar :: Vector3 GLfloat -> (GLfloat, GLfloat)

Given an Vector3 x y z or Vertex3 x y z | Convert Cartesian Coordinates to Polar Coordinates | beta x-y plane, | alpha x-z plane | | y z | | / | ---|/--- x | /

roty :: Floating a => a -> [[a]]

right hand coordinates system in OpenGL | | x-z plane

rotz :: Floating a => a -> [[a]]

x-y plane

drawCircleVec :: GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat -> IO ()

Given an radius r, center Vertex3 vx, and Vector ve Draw circle(r, vx) which is perpendicular to ve |

matVx :: [[a]] -> Vertex3 a

matrix to Vertex

matVe :: [[a]] -> Vector3 a

matrix to Vector

veMat :: Vector3 a -> [[a]]

Vector to matrix

drawCircle' :: Vertex3 GLfloat -> Double -> IO ()

draw circle with center and radius | ceneter radius drawCircle::(GLfloat, GLfloat, GLfloat) -> Double -> IO() drawCircle cen r = drawPrimitive LineLoop red $ circle cen r

drawCircleColor :: Vertex3 GLfloat -> Color3 GLdouble -> Double -> IO ()

drawCircleColor' :: Color3 GLdouble -> Double -> Vertex3 GLfloat -> IO ()

conic :: [(GLfloat, GLfloat, GLfloat)]

Conic Parameter Equation | | gx http://localhost/html/indexThebeautyofTorus.html |

ray :: Vertex3 GLfloat -> GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat

r(t) = p0 + t(p1 - p0)

det2 :: Num a => [[a]] -> a

determinant of two dimension matrix

inverseR :: [[Rational]] -> [[Rational]]

inv2 :: Fractional a => [[a]] -> [[a]]

inverse of two dismension matrix [a b] inv 1/det [d -b] [c d] [-c a]

intersectLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 GLfloat, [[GLfloat]])

data Seg a = Seg a a | fun::Seg GLflat -> Seg GLfloat -> Maybe(Vertex3 GLfloat, [[GLfloat]]) | fun (Seg (Vertex3 x0 y0 z0) (Vertex3 x1 y1 z1)) = Nothing

find the intersection of two line | if p0 p1 and q0 q1 are not parallel, | return the vertex of intersection and s and t | gx http://localhost/pdf/intersectionLine.pdf

let inter = let p0 = Vertex3 (x0=0) (y0=0) (z0=0) p1 = Vertex3 (x1=1) (y1=1) (z1=0) q0 = Vertex3 (a0=2) (b0=0) (c0=0) q1 = Vertex3 (a1=2) (b1=1) (c1=0) in intersectLine p0 p1 q0 q1

intersectLine' :: [Vertex3 GLfloat] -> Maybe (Vertex3 GLfloat, [[GLfloat]])

Four Vertex3 in a list

pointToLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

TODO: fix, the code only handles two dimension. | Given a point: p0, line: q0 q1 | Compute the the distance from p0 to line: q0 q1 |

intersectedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat)

find the intersection of two segments, otherwise return Nothing

isIntersectedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Bool

intersectLineR :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational, [[Rational]])

perpPlane :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Given a point p0 and three pts: q0, q1, q2, | Compute the intersection of line is perpendicular to the plane and passes point p0

If three pts (q0, q1, q2) are colinear, return Nothing otherwise, return Just (Vertex3 GLfloat)

lineIntersectPlane :: Vertex3 GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Line intersects Plane | Given Line: p0 p1 and Plane: q0 q1 q2 | Compute the intersection of line and plane | p1 = p0 +: h(p1 -: p0)

lineInterpolate :: (GLfloat, GLfloat, GLfloat) -> (GLfloat, GLfloat, GLfloat) -> [(GLfloat, GLfloat, GLfloat)]

draw line from two points | gx UserscatmyfilebitbucketmathdrawLine.pdf | ----------------------------------------------------------------------------- | p &= (1-t) p0 - t p_1 | ----------------------------------------------------------------------------

convexHull :: Int -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat] -> [[Vertex3 GLfloat]]

naive algorithm to compute the convex hull | convexHull (length pts) y-Axis_top_pts y-Axis_top_pts smallest_angle_(-1, 0)_and_vec(x0 x1) all_pts | need more explanation |

projv :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

projection from u onto v in Vector3

nr :: Vector3 GLfloat -> GLfloat

norm of vector

uv :: Vector3 GLfloat -> Vector3 GLfloat

unit normal of vector

rod :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat -> Vector3 GLfloat

Rodrigue formula | | u rotates around v in angle θ in right hand rule | | rejection = u - p | | gx http://localhost/pdf/rotate_arbitrary_axis.pdf |

cosVec2 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Float

cosVec2 p0 p1 p2 | vec(p1->p0) dot vec(p1->p2) = |vec(p1->p0)||vec(p1->p2) cos(beta) |

Orphan instances

Num a => Num (Vector3 a)
(Vector3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vector3 a -> Vector3 a -> Vector3 a

(-) :: Vector3 a -> Vector3 a -> Vector3 a

(*) :: Vector3 a -> Vector3 a -> Vector3 a

negate :: Vector3 a -> Vector3 a

abs :: Vector3 a -> Vector3 a

signum :: Vector3 a -> Vector3 a

fromInteger :: Integer -> Vector3 a

Num a => Num (Vertex3 a)
(Vertex3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(-) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(*) :: Vertex3 a -> Vertex3 a -> Vertex3 a

negate :: Vertex3 a -> Vertex3 a

abs :: Vertex3 a -> Vertex3 a

signum :: Vertex3 a -> Vertex3 a

fromInteger :: Integer -> Vertex3 a