AronGraphic

Documentation

lightDiffuse :: Color4 GLfloat

lightAmbient :: Color4 GLfloat

lightPosition :: Vertex4 GLfloat

lightSpecular :: Color4 GLfloat

data SegEndPt

Constructors

No

End

Beg

Both

Cen

All

data NormalDir

counter clockwise | clock wise

Constructors

NCCW

NCW

dist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute the distance between two points

\( v = (x, y, z) \)

\( \| v \| = \sqrt{ x^2 + y^2 + z^2} = \sqrt{ v \cdot v} \)

let v1 = Vertex3 1 2 3
let v2 = Vertex3 2 3 4
dist v1 v2

sqdist :: Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute the norm-squared

\( v = (x, y, z)\)

\( |v|^2 = x^2 + y^2 + z^2 \)

let v1 = Vertex3 1 2 3
let v2 = Vertex3 2 3 4
sqdist v1 v2

dot3ve :: Num a => Vector3 a -> Vector3 a -> a

dot product for Vector3

dot3vx :: Num a => Vertex3 a -> Vertex3 a -> a

dot product for Vertex3

(+:) :: Num a => Vertex3 a -> Vector3 a -> Vertex3 a

(-:) :: Num a => Vertex3 a -> Vertex3 a -> Vector3 a

(⊥) :: Vector3 GLfloat -> Vector3 GLfloat

perpcw :: Num a => Vector3 a -> Vector3 a

perpccw :: Num a => Vector3 a -> Vector3 a

(*:) :: Num a => a -> Vector3 a -> Vector3 a

(**:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(*>:) :: GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

(>>:) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

(.>) :: Int

(⊙) :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat

dot product odot

(∈) :: (Num a, Ord a) => a -> [a] -> Bool

(∘) :: Num a => a -> Vector3 a -> Vector3 a

(∑) :: Int

(∎) :: GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

v2x :: Num a => Vector3 a -> Vertex3 a

x2v :: Num a => Vertex3 a -> Vector3 a

red :: Color3 GLdouble

Shorten realToFrac rf::(Real a, Fractional b) => a -> b rf = realToFrac

------------------------------------------------------------------------------- | Fri Dec 7 14:35:38 2018 | three colors: data Color3 a = Color3 !a !a !a

green :: Color3 GLdouble

blue :: Color3 GLdouble

skew :: Num a => Vertex3 a -> [[a]]

Form a skew symmetric matrix from \( \color{red}{Vertex3} \)

cross product of a x b = [a] b where [a] is skew symmetric matrix from vector a colIndex x rowIndex = [c][r]

Vertex3 a1 a2 a3

\[ \begin{bmatrix} 0 & -a_3 & a_2 \\ a_3 & 0 & -a_1 \\ -a_2 & a_1 & 0 \\ \end{bmatrix} \]

skew' :: Num a => Vector3 a -> [[a]]

Form a skew symmetric matrix from \( \color{red}{Vector3} \)

cross product of a x b = [a] b where [a] is skew symmetric matrix from vector a colIndex x rowIndex = [c][r]

Vertex3 a1 a2 a3

\[ \begin{bmatrix} 0 & -a_3 & a_2 \\ a_3 & 0 & -a_1 \\ -a_2 & a_1 & 0 \\ \end{bmatrix} \]

(⊗) :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

Cross product of two vectors.

The directin is determinated by the the Right Hadle Rule

isCoplanar :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

check whether four points are coplanar

| Userscatmyfilebitbucketmathcoplanar.pdf

cross :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

normal3 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vector3 GLfloat)

compute the normal of three points

normal3' :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

compute the normal of plane with two given vectors

vec :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat

Vector: p0 -> p1 = p1 - p0

ray :: Vertex3 GLfloat -> GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat

Given a point and vector,

| draw a line passes the point and alone the vector

| r(t) = p0 + t(p1 - p0)

perpenLine :: GLfloat -> GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat]

Given s, t, p0 and p1 1. c₀ = center p₀ p₁ 2. p1' = p0 + s(⊥ p0 p1), p2' = p0 + t(⊥ p0 p1) 3. draw line from p1' to p2'

curvePt :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> [Vertex3 GLfloat]

drawCurve :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> Color3 GLdouble -> IO ()

Given a function \(f\), interval \( (a, b) \)

draw the curve from \(a\) to \(b\)

mapM_ (\n -> drawCurve (\x -> x^n) (-1.0, 1.0) green) [1..20] 

drawSurface :: (GLfloat -> GLfloat -> GLfloat) -> IO ()

draw Surface for equation \( f(x, y) = x^2 + y^2 \) form

Draw \( f (x, y) = x^2 + y^2 \)

drawSurface (\x y -> x^2 + y^2)

plotPts :: Color3 GLdouble -> [GLfloat] -> IO ()

Plot all pts

Plot_Points

interval from [-len/2.. len/2]
let pts = [0.3, 0.4, 0.1, 0.2] 
plotPts red pts 
plotPts green $ quickSort1 pts 

tangentLine :: Fractional a => (a -> a) -> (a, a) -> a -> [Vertex3 a]

Draw a line that is tangent at pt (c, f(c)) where x in [x0, x1]

tangentLine

Given a function f, and x0 on x-Axis

Derive a tangent line at (c, f(c)) with slop = f'(c)

Interpolate (x0, f(x0)) and (x1, f(x1))

interpolate' :: Fractional a => Integer -> Vertex3 a -> Vertex3 a -> [Vertex3 a]

drawTangentLine :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> Color3 GLdouble -> IO ()

Given a function f, interval (a, b),

draw a tangent line at (x0, f(x0)) from a to b >f::x -> y

(a, b) is interval

differentiate at (x0, f x0)

normalLine :: NormalDir -> (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> [Vertex3 GLfloat]

NormalDir is either counter clockwise or clockwise for normal

(x0, x1) is interval for the normal f is any function x=c is the tangent point at (c, f c)

data NormalDir = NCCW | NCW deriving (Eq)

normalLineNew :: NormalDir -> (GLfloat -> GLfloat) -> GLfloat -> GLfloat -> [Vertex3 GLfloat]

drawNormalLine :: (GLfloat -> GLfloat) -> (GLfloat, GLfloat) -> GLfloat -> Color3 GLdouble -> IO ()

drawNormalLineNew :: (GLfloat -> GLfloat) -> GLfloat -> GLfloat -> Color3 GLdouble -> IO ()

cen :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat

compute the center of two vertices/vertex

isPerpen :: Vector3 GLfloat -> Vector3 GLfloat -> Bool

check two vectors whether they are perpendicular

isColinear :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Bool

Check whether three pts are colinear

NOTE

If two points or three points are overlapped, then they are still colinear

   f(t) = p0 + t(p1 - p0)
   f(s) = p0 + s(p2 - p0)
   f(t) = f(s) =>
   t(p1 - p0) = s(p2 - p0)
   t v1 = s v2, let A = [v1 v2]
   det A =? 0
   det A = 0 => p0 p1 p2 are colinear
   

isInSegment :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe Bool

Check whether a given point is inside the segment

NOTE use ptOnSegment instead

Assume three pts are different pts

Given p0, q0 q1, check whether p0 is inside the segment of q0 q1

If they are colinear

then check the distance \( \overline{p_0 q_0} \) \( \overline{p_0 q_1} \) and \( \overline{q_0 q_1} \)

Using following Affine combination on points formula

Linear combination on vector \( \{ x_1, x_2, \dots \} \)

\[ \begin{aligned} \sum_{i=1}^{n} \alpha_{i} x_i \end{aligned} \]

If the sum of all the coefficients is 1 \[ \begin{aligned} \sum_{i=1}^{n} \alpha_{i} &= 1 \end{aligned} \] is called Affine Combination of \( \{x_1, x_2, \dots \} \)

NOTE If we extend above definition to highter power on \(t\), \(t\) can have any power \( n > 0 \) or \( t^{n} \)

\( [(1 - t) + t]^{n} \) is just Binomial Expansion

Bezier_Curve

\[ \begin{aligned} 1 &= [(1 - t) + t]^{1} \\ Q &= (1 - t)p_0 + t p_1 \\ \\ 1 &= [(1 - t) + t]^{2} = (1 - t)^{2} + 2t(1-t) + t^2 \\ Q &= (1-t)^{2} p_0 + 2t(1-t) p_1 + t^2 p_2 \\ \end{aligned} \]

In out case, there are only two points \( p_0, p_1 \) so \( p \) is the Affine Combination of \( p_0, p_1 \) \[ \begin{aligned} p &= (1 - t) p_0 + t p_1 \\ \end{aligned} \] If \( 0 < t < 1 \) then the point on the segment \( \overline{p_0 p_1} \) but not overlapping on \( p_0, p_1 \)

If \( t = 0 \) or \( t = 1 \) then the point \(p\) is overlapping \( p_0 \) or \( p_1 \)

If \( t < 0 \) or \( t > 1 \) then the point \(p\) is NOT on the segment \( \overline{p_0 p_1} \)

p0 = Vertex3 0.5 0.5 0
q0 = Vertex3 1 1 0
q1 = Vertex3 0 0 0
isInSegment p0 q0 q1
Just True

data PtSeg

Constructors

OnEndPt

InSeg

OutSeg

ptOnSegment :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> PtSeg

Better version of isInSegment

If \(p_0\) is overlapped with \(q_0, q_1\) then OnPt

If \(\overline{p_0 q_0} + \overline{p_0 q_1} > 0\) then OutSeg

Else InSeg

   data PtSeg = OnEndPt -- ^ Overlapped pt
                | InSeg   -- ^ Inside the segment
                | OutSeg  -- ^ Out the segment
   

Maybe use better name: crossSegments - No endpt is overlapped

1. Three pts are colinear
    1. one endpts is overlapped
    2. No endpt is overlapped, one endpt is "inside" the segment
    3. Not intersected
2. Four pts are colinear
     1. No endpts is overlapped
         1. one segment "inside" the other segment
         2. one segment "outside" the other segment
     2. One endpts is overlapped
         1. one endpt is "outside" a segment
         2. one endpt is "inside" a segment
     3. two endpts are overlapped => same segment
3. No three pts are colinear
    1. If two segment is intersected, one must cross other segment
    2. Two segments DO NOT intersect

drawPrimitive :: PrimitiveMode -> Color3 GLdouble -> [(GLfloat, GLfloat, GLfloat)] -> IO ()

draw primitive such lines, points, lineloop etc

PrimitiveMode

   PrimitiveMode
   Lines
   LineStrip
   LineLoop
   TriangleStrip
   TriangleFan
   Quad
   QuadStrip
   Polygon
   Patches
   

drawPrimitive' :: PrimitiveMode -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

draw primitive such lines, points, lineloop etc

PrimitiveMode

   Lines
   LineStrip
   LineLoop
   TriangleStrip
   TriangleFan
   Quad
   QuadStrip
   Polygon
   Patches
   

randomVertex

list <- randomVertex 60
drawPrimitive' Lines red list

drawLines :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegmentWithEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

Draw Segment with Two End Pts

v = [Vertex3 0.1 0.1 0.2, Vertex3 0.4 0.2 0.7]
drawSegmentWithEndPt red v

v = [[Vertex3 0.1 0.1 0.2, Vertex3 0.4 0.2 0.7]]
mapM_(\x -> drawSegmentWithEndPt red x) vv

drawSegmentNoEndPt :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

drawSegments :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

Draw a set of segments

Segment contains two end points, one is begin point, other is end point

[p0, p1] = p0 -> p1

drawSegment :: Color3 GLdouble -> (Vertex3 GLfloat, Vertex3 GLfloat) -> IO ()

Draw one segment from p0 to p1 | drawSegment color (p0, p1) |

drawSegmentNoEnd :: Color3 GLdouble -> (Vertex3 GLfloat, Vertex3 GLfloat) -> IO ()

Draw one segment with no endpt

drawSegment' :: SegEndPt -> Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

definition: Segment contains two end points, one is begin point, other is end point [p0, p1] = p0 -> p1

data SegEndPt = No -- no pt, just a segment | End -- end pt | Beg -- begin pt | Both -- begin and end pts | Cen -- center pt | All -- all pts: begin, end and ceneter

drawSegmentArg :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

Draw intersection pt from two segments | Two segments need not have an intersection.

drawSpherePt :: Vertex3 GLfloat -> GLfloat -> [Vertex3 GLfloat]

draw sphere center at c with radius r

β is in x-y plane, it rotates around z-Axis. ϕ is x-z plane, it rotates aroound y-Axis

\( r*\cos(ϕ) \) is the radius of circle cut through x-y plane

drawPlane :: Color3 GLdouble -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> IO ()

draw plane with three points | no check whether they are co-linear

circlePt :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

Generate a circle

let radius = 0.1
let pts = circlePt (Vertex3 0 0 0) radius 

circle' :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

\(\color{red}{Deprecated} \) Use circlePt

Fri Feb 15 11:13:17 2019

Draw xy-plane circle

circleXY :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

Draw xz-plane circle

circleXZ :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

Draw xz-plane circle

circleYZ :: Vertex3 GLfloat -> Double -> [Vertex3 GLfloat]

Draw yz-plane circle

drawDot :: Vertex3 GLfloat -> IO ()

draw dot, small circle

cartToPolar :: Vector3 GLfloat -> (GLfloat, GLfloat)

Given an Vector3 x y z or Vertex3 x y z | Convert Cartesian Coordinates to Polar Coordinates | beta x-y plane, | alpha x-z plane | | y z | | / | ---|/--- x | /

roty :: Floating a => a -> [[a]]

right hand coordinates system in OpenGL | | x-z plane

rotz :: Floating a => a -> [[a]]

x-y plane

coordTip :: IO ()

coordTipX :: IO ()

coordTipY :: IO ()

coordTipZ :: IO ()

renderCoordinates :: IO ()

Coordinate with tips

positive dir is the tips dir

drawCircleVec :: GLfloat -> Vertex3 GLfloat -> Vector3 GLfloat -> IO ()

Given an radius r, center Vertex3 vx, and Vector ve Draw circle(r, vx) which is perpendicular to ve |

matVx :: [[a]] -> Vertex3 a

matrix to Vertex

matVe :: [[a]] -> Vector3 a

matrix to Vector

veMat :: Vector3 a -> [[a]]

Vector to matrix

drawRect :: (Vertex3 GLfloat, Vertex3 GLfloat) -> IO ()

drawQuads :: [Vertex3 GLfloat] -> IO ()

drawCircle' :: Vertex3 GLfloat -> Double -> IO ()

draw circle with center and radius

drawCircle cen r = drawPrimitive LineLoop red $ circle cen r

drawCircleXYZ :: Vertex3 GLfloat -> Double -> IO ()

drawCircleXYZColor :: Vertex3 GLfloat -> Double -> Color3 GLdouble -> IO ()

drawCircleColor :: Vertex3 GLfloat -> Color3 GLdouble -> Double -> IO ()

draw circle with center , Color3, radius

drawCircleColor (Vertex3 0.1 0.2 0.3) red 0.5

drawCircleColor' :: Color3 GLdouble -> Double -> Vertex3 GLfloat -> IO ()

Similar to drawCircleColor, but it can do more

draw two circles with different centers

mapM_ (drawCircleColor' red 0.5) [Vertex3 0.1 0.2 0.3, Vertex3 0.2 0.3 04]

conic :: [(GLfloat, GLfloat, GLfloat)]

Conic Parameter Equation

gx Conic

det2 :: Num a => [[a]] -> a

determinant of two dimension matrix

inverseR :: [[Rational]] -> [[Rational]]

inv2 :: Fractional a => [[a]] -> [[a]]

Inverse of two dimension matrix

Inverse_matrix_determinant

\[ \begin{equation} \begin{aligned} A &= \begin{bmatrix} a & b \\ c & d \\ \end{bmatrix} \\ A^{ -1} &= \frac{1}{ \det A } \begin{bmatrix} d & -b \\ -c & a \\ \end{bmatrix} \end{aligned} \end{equation} \] * inverse should be used in general.

Or isInver should be used for large matrix because isInver is implemented in QR_Decompoisition

   isInver::(Fractional a, Ord a)=> [[a]] -> Bool
   isInver m = if len (filter(cx) 0 then False else True
   

Following function is implemented in Gaussian Elimination

There is some Integer overflow issue, it only works for small matrix, e.g. 10 by 10

   isInvertible::[[Integer]]->Bool
   

data SegColinear

Constructors

Colinear3
If three pts are colinear => Colinear3

Colinear4
If four pts are colinear => Colinear4

None

fourPtColinear :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> SegColinear

If four points are colinear then return Colinear4

If only three points are colinear then return Colinear3

Else return None

data SegColinear = Colinear3 -- ^ If three pts are colinear => Colinear3
                   | Colinear4 -- ^ If four pts are colinear => Colinear4
                   | None deriving (Eq, Show) -- ^ else => None

The function uses isColinear

intersectLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 GLfloat, [[GLfloat]])

Find the intersection of two lines, also see intersectSeg

Assume that two endpoints of each segment are not overlapped.(segment has non-zero length)

NOTE This function is ONLY for two dimensions

Line extends infinitly in both direction

If both line intersect at one pt => return the pt 1. Intersection can be on a segments or NOT on a segment

If four points are colinear: return Nothing, see fourPtColinear

let p0 = Vertex3 0 0 0
let p1 = Vertex3 1 0 0
let q0 = Vertex3 1 0 0
let q1 = Vertex3 2 0 0
intersectLine p0 p1 q0 q1
Nothing

two dimensions determinant is used here

The Intersection of two line is NOT necessary in their segments

If two lines are parallel or overlapped then return Nothing

Else return the intersection and \( s, t \)

Given four pts: \(p_0, p_1\) and \(q_0, q_1\) \[ \begin{aligned} v_0 &= p_1 - p_0 \\ v_1 &= q_1 - q_0 \\ A &= \begin{bmatrix} v_0 & v_1 \end{bmatrix} \\ \det A &= 0 \quad \text{they are linearly dependent } \end{aligned} \]

   data Seg a = Seg a a
   fun::Seg GLflat -> Seg GLfloat -> Maybe(Vertex3 GLfloat, [[GLfloat]])
   fun (Seg (Vertex3 x0 y0 z0) (Vertex3 x1 y1 z1)) = Nothing
   

intersect_line

   intersectLine p0 p1 q0 q1
   (Just ((Vertex3 2.0 2.0 0), [[2.0],[2.0]]))= intersectLine
                                               (Vertex3 0 0 0)
                                               (Vertex3 1 1 0)
                                               (Vertex3 2 0 0)
                                               (Vertex3 2 1 0)
   

intersectLine2 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 GLfloat, (GLfloat, GLfloat))

If two line parallel or four pts colinear => return Nothing

Else there is intersection pt, pt maybe on on segment or NOT on segment

Sun Feb 17 15:32:59 2019 This function should replace intersectLine

onePtOverlappedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat, (GLfloat, GLfloat))

Given two segments: \( (p_0, p_1), (q_0, q_1) \), find the overlapped endPt

Precondition: Four pts are NOT colinear \( \Rightarrow \) any three pts are NOT colinear

If two segments are overlapped at one endPt, return Maybe(Vertex3 GLfloat, GLfloat s, GLfloat t) else return Nothing

Four cases

\( p_0 = q_0 \)

\( p_0 = q_1 \)

\( p_1 = q_0 \)

\( p_1 = q_1 \)

onept_overlapped_segment

intersectLine' :: [Vertex3 GLfloat] -> Maybe (Vertex3 GLfloat, [[GLfloat]])

Four Vertex3 in a list

pointToLine :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

TODO: fix, the code only handles two dimension. | Given a point: p0, line: q0 q1 | Compute the distance from p0 to line: q0 q1 |

intersectSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat)

Find the intersection of two segments, intersectLine or intersectSegNoEndPt

Assume that two endpoints of each segment are not overlapped.(segment has non-zero length)

If two EndPts from different segments are over overlapped => then Maybe(Vertex3 endpt)

NOTE This function is ONLY for two dimensions

NOTE endpoins of different segments may be coincide

intersectSeg (p0, p1) (q0, q1)
Nothing

intersectSeg (p0, p1) (q0, q1)
Vertex3 x y z

The function is based on intersectLine * If two segments are parallel or overlapped then return Nothing

TODO: add test cases

intersectSegNoEndPt :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat)

intersection excluding two EndPts, intersectLine or intersectSeg

Sun Feb 17 19:24:22 2019

There are some issues in endpts overlapped

Deprecated, Should use intersectSegNoEndPt2

If four pts are colinear \( \Rightarrow \) Nothing

intersectSegNoEndPt2 :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Maybe (Vertex3 GLfloat)

intersection excluding two EndPts, intersectLine or intersectSeg

Sun Feb 17 19:26:05 2019

Fixed some bugs from intersectSegNoEndPt

isIntersectedSeg :: (Vertex3 GLfloat, Vertex3 GLfloat) -> (Vertex3 GLfloat, Vertex3 GLfloat) -> Bool

If four pts are colinear return False

If the intersection is within \( s \in [0.0, 1.0], t \in [0.0, 1.0] \) return True

Else return False

intersectLineR :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational, [[Rational]])

intersect line, return Rational

threePtDeterminant :: Fractional a => Vertex3 a -> Vertex3 a -> Vertex3 a -> a

Three points rotation order can be determinated by the \( \color{red}{\text{Right Hand Rule}} \)

Three_Points_Direction

If given three points in following order:

p₀ = (1, 0)

p₁ = (0, 0)

p₂ = (1, 0)

then vectors are computed in following:

v10 = p₀ - p₁

v12 = p₂ - p₁

matrix can be formed as following:

m = [v10, v12]

\( \vec{v_{01}} \times \vec{v_{02}} \) in Right Hand Rule

\[ m = \begin{bmatrix} 1 & 0 \\ 0 & 1 \\ \end{bmatrix} \\ \det m = 1 > 0 \\ \]

If the three points are collinear:

\( \det M = 0 \)

If the order of three points in clockwise order:

\( \det M > 0 \)

If the order of three points in counter clockwise order:

\( \det M < 0 \)

threePtCCW :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points in Counter ClockWise order

\( \det M > 0 \)

threePtCW :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points in ClockWise order

\( \det M < 0 \)

threePtCollinear :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Three points are collinear

\[ \begin{aligned} p_0 &= (x_0, y_0) \\ p_1 &= (x_1, y_1) \\ p_2 &= (x_2, y_2) \\ u &= p_0 - p_1 \\ u &= (x_0 - x_1, y_0 - y_1) \\ v &= p_2 - p_1 \\ v &= (x_2 - x_1, y_2 - y_1) \\ M &= \begin{bmatrix} u & v \end{bmatrix} \\ &\mbox{If three pts are colinear, then } \\ \det M &= 0 \\ \end{aligned} \]

ptOnLeftLine :: (Fractional a, Ord a) => Vertex3 a -> Vertex3 a -> Vertex3 a -> Bool

Given a point \( p_0 \), a line \(q_0, q_2 \) === Check whether a point \(p_0\) is on the left size of a line \( q_0, q_2 \), it is based on threePtDeterminant with additional condtion: \(y_2 > y_0 \) or \( y_2 < y_0 \) since

NOTE y2 \(\neq\) y0 where given (Vertex3 x0 y0 z0) and (Vertex3 x2 y2 z2)

Assume \( p_0, p_2 \) are not coincide

If pt on the line, return False

If pt on the right side of the line, return False

Else return True

\[ \begin{aligned} p_1 &= (x_1, y_1) \\ p_0 &= (x_0, y_0) \\ p_2 &= (x_2, y_2) \\ u &= \overrightarrow{p1 p0} \\ v &= \overrightarrow{p1 p2} \\ \end{aligned} \]

if \( y_2 > y_0 \) form a determinant \[ \begin{aligned} \det \begin{vmatrix} u & v \end{vmatrix} > 0 \end{aligned} \] if \( y_2 < y_0 \) form a determinant \[ \begin{aligned} \det \begin{vmatrix} u & v \end{vmatrix} < 0 \end{aligned} \]

If \( y_2 = y_0 \), compare \( x_0, x_2 \)

Check whether a point is below a line \(p_0, p_2\)

if \( x_2 < x_0 \) \[ \begin{aligned} \det \begin{vmatrix} u & v \end{vmatrix} > 0 \end{aligned} \]

if \( x_2 > x_0 \) \[ \begin{aligned} \det \begin{vmatrix} u & v \end{vmatrix} < 0 \end{aligned} \]

ptInsideTri :: Vertex3 GLfloat -> (Vertex3 GLfloat, Vertex3 GLfloat, Vertex3 GLfloat) -> (Bool, GLfloat)

Check whether a point \(p_0\) is inside a triangle \( \triangle ABC \)

NOTE the order of three pts: \(A, B, C\) does't matter. e.g CCW or CW

\(p_0\) can NOT be the same point as \(A, B, C\)

If \(p_0\) is collinear with AB, BC, or AC, then \(p_0\) is considered inside the \( \triangle ABC \)

\(p_0\) is the point that is tested three points \(A, B, C\) forms a triangle

if point p0 inside the triangle, return true

The sum of three angles are in degree.

let p0 = Vertex3 0.1 0.1 0
let a  = Vertex3 1 0 0
let b  = Vertex3 0 0 0
let c  = Vertex3 0 1 0

ptInsideTri p0 a b c
(True, 360.0)

TODO: how to check whether a pt is inside a n-polygon ptInsidePolygon N-Polygon

perpPlane :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Given a point p0 and three pts: q0, q1, q2,

Compute the intersection of line is perpendicular to the plane and passes point p0

If three pts (q0, q1, q2) are colinear, return Nothing

otherwise, return Just (Vertex3 GLfloat)

lineIntersectPlane :: Vertex3 GLfloat -> Vector3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> Maybe (Vertex3 Rational)

Line intersects Plane | Given Line: p0 p1 and Plane: q0 q1 q2 | Compute the intersection of line and plane | p1 = p0 +: h(p1 -: p0)

convexHull :: Int -> [Vertex3 GLfloat] -> [[Vertex3 GLfloat]]

Slow Algorithm 1 \( \color{red}{ \mathcal{O}(n^2)} \)

find the a point \( p_0 \) which has the maximum y-axis value

find the a point \( p' \) from \( p_0 \) so that the rest of all the points on one side of the segment \( \overline{p_0 p'} \)

continua with the rest of points

Slow_Algorithm

p0 p1 p2 p3
(p0, p1) (p0, p2) (p0, p3)
(p1, p2) (p1, p3)
(p2, p3)

Naive Algorithm to compute the convex hull in \( \color{red}{\mathcal{O}(nh)} \) where \( \color{red}{n} \) is the number of points on the plane, \( \color{red}{h} \) is the number of points on the convex hull

If all the points are around the convex hull, then the runtime will be \( \color{red}{\mathcal{O}(n^2)} \)

convexHull n pts
n is the total of vertices
pts is length pts

p0 = Vertex3 0 0 0
p1 = Vertex3 1 0 0
p2 = Vertex3 0 1 0
p3 = Vertex3 0.1 0.1 0
pts= [p0, p1, p2, p3]
n  = 4 = len pts
pts = [p0, p1, p2, p3]
exp= [[p2, p1],
      [p1, p0],
      [p0, p2]]

Algorithm ConvexHull

Find a top vertex: B(Vertex3 x y z) with maximum y-Axis value

Construct a new vertex: A(Vertex3 x+1 y z) = top(Vertex3 x + 1 y z) according to top(Vertex3 x y z)

Compute the angle \( \cos ∠ABC \) from three points: \(A, B, C\)

If there are only two points in pts for the list of Vertex3 e.g. p0 p1

then segments: \(\overline{AB} \) and \(\overline{BC} \) are drawn Three_pts_angle

n = length pts \(\Rightarrow\) n vertices \(\Rightarrow\) n edges. if there is two vertices, then \((p_0, p_1)\) and \((p_1, p_0)\)

vec(top x1) and vec(top x2)

\(\vec{v_1}\) = top -> x1 = x1 -: top

\(\vec{v_2}\) = top -> x2 = x2 -: top

TODO: Handle the case where the three points are colinear.

TODO: If there are more than one point that have maximum y-axis, then choose the left most point Top_Point

\( \color{red}{TODO} \): use eliminate points technic to improve the algorithm Eliminate_Points

It seems that the algorithm is still \( \color{red}{ \mathcal{O}(nh)} \) Given n vertices on a plane. There are \(h\) points on the convex hull. Assume \( \frac{1}{h} n \) vertices may be eliminated when the next point is found on the convex hull. The total number of steps are: \[ \begin{aligned} s &= n + \frac{h-1}{h} n + \frac{h-2}{h} n + \dots + \frac{1}{h} n \\ s &= n h (1 + 2 + 3 + \dots + h) \\ s &= n h \frac{(1 + h)h}{2} \\ s &= n \frac{1+ h}{2} \\ &= \mathcal{O}(nh) \\ \end{aligned} \]

When the algo walks around the convex hull in counter clokcwise order.

segment can be drawn from top point to previous point

All the points can be removed if they are on the left side of the segment

\( \color{red}{TODO} \) use other algo: check whether a point is on the left or right side of a segment

The algo can be check with the determinant of two vectors threePtDeterminant Three_Points_Direction

convexHull2 :: Int -> [Vertex3 GLfloat] -> [(Vertex3 GLfloat, Vertex3 GLfloat)]

Same as convexHull, return [(Vertex3 GLfloat, Vertex3 GLfloat)]

convexHullAllSeg :: [Vertex3 GLfloat] -> [(Vertex3 GLfloat, Vertex3 GLfloat)]

Connect all pts inside convex hull without crossing segments

All_Segments_ConvexHull

   let vexList = [Vertex3 0.1 0.1 0.1, Vertex3 0.4 0.3 0.6, Vertex3 0.6 0.8 0.2]
   mapM_(vx -> do 
               drawSegment green vx 
               threadDelay 500
               ) $ convexHullAllSeg vexList
   

ptInsidePolygon :: Vertex3 GLfloat -> [Vertex3 GLfloat] -> Bool

Check whether a pt is inside a polygon (assume the polygon is convex)

This is essentially a Covex Hull problem => use Convex Hull Algorithm

TODO: Add test cases, never test it yet

containPt :: Vertex3 GLfloat -> [Vertex3 GLfloat] -> Bool

Given a \(p_0\): Vertex3 and [Vertex3], check whether the list contain p0

projv :: Vector3 GLfloat -> Vector3 GLfloat -> Vector3 GLfloat

Projection from \(u\) onto \(v\) in Vector3

projection

nr :: Vector3 GLfloat -> GLfloat

Compute the norm of a vector

\(v = (x, y, z)\)

\(|v| = \sqrt{ x^2 + y^2 + z^2} \)

uv :: Vector3 GLfloat -> Vector3 GLfloat

Normalize a vector

e.g \( \|\vec{v}\| = 1 \)

rod :: Vector3 GLfloat -> Vector3 GLfloat -> GLfloat -> Vector3 GLfloat

Rodrigue formula

\(u\) rotates around \(v\) in angle \(\phi\) in right hand rule

rejection = \(u - p\)

gx Rotation

cosVex3 :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> GLfloat

Compute an angle from three points: \(a, b, c \) with dot product

let a = Vertex3 1 0 0
let b = Vertex3 0 0 0
let c = Vertex3 1 0 0
cosVex3 a b c
0.0

\( \vec{ba} = a - b \)

\( \vec{bc} = c - b \)

\[ \begin{equation} \begin{aligned} \vec{ba} \circ \vec{bc} &= | \vec{ba} | | \vec{bc} | \cos{\angle ABC} \\ \cos{\angle ABC} &= \frac{ \vec{bc} \circ \vec{bc} }{| \vec{ba} | | \vec{bc}|} \\ \end{aligned} \end{equation} \]

nonCrossSegmentNoEndPt :: [(Vertex3 GLfloat, Vertex3 GLfloat)] -> Vertex3 GLfloat -> [(Vertex3 GLfloat, Vertex3 GLfloat)]

Find all the segments DO NOT Cross some new segments

Current algorithm is brute force

NOTE use Convex Hull algo might be faster

NOTE EnPts are excluded, please see intersectSegNoEndPt

Given a list of segments and a point

non_cross_segment

non_cross_segment_2

Given Segments \( [(B, E), (E, A), (A, D), (D, C)] \) pt: \(F\)

return => \( [(A, F), (F, D), (F, C)] \)

   let p0 = Vertex3 0 0 0
       p1 = Vertex3 0.5 0 0
       p2 = Vertex3 0 0.5 0
       q0 = Vertex3 1 1 0
       ls = [(p0, p1), (p0, p2)]
       exp= sort [(p0, q0), (p1, q0), (p2, q0)]
       in exp == (sort $ nonCrossSegmentNoEndPt ls q0)

   let p0 = Vertex3 0 0 0
       p1 = Vertex3 0.5 0 0
       p2 = Vertex3 0 0.5 0
       q0 = Vertex3 1 1 0
       ls = [(p0, p1), (p1, p2)]
       exp= sort [(p1, q0), (p2, q0)]
       in exp == (sort $ nonCrossSegmentNoEndPt ls q0)
   

remove all the segments with same vertex from old segments list then check all the new segments with the rest of old segments list

rmAdjacentSeg :: Vertex3 GLfloat -> [(Vertex3 GLfloat, Vertex3 GLfloat)] -> [(Vertex3 GLfloat, Vertex3 GLfloat)]

Remove all adjacent edges or adjacent segments from a list of segments/edges

affineTri :: Vertex3 GLfloat -> Vertex3 GLfloat -> Vertex3 GLfloat -> [Vertex3 GLfloat]

Affine Combination on three points

Draw all the points inside a triangle p0 p1 p2

The number of steps is \( n = 10 \)

Extend it to polygon

Affine_Triangle

drawSegmentFromTo :: Color3 GLdouble -> [Vertex3 GLfloat] -> IO ()

Draw all segments from Vertex3 a To Vertex3 b

grid :: [[Vertex3 GLfloat]]

2d grid on x-y-plane, z=0

draw grid

mapM_ (\row -> drawSegmentFromTo red row ) grid 
mapM_ (\row -> drawSegmentFromTo red row ) $ tran grid 

grid2 :: (GLfloat -> GLfloat -> GLfloat) -> [[Vertex3 GLfloat]]

Orphan instances

Num a => Num (Vector3 a)
(Vector3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vector3 a -> Vector3 a -> Vector3 a

(-) :: Vector3 a -> Vector3 a -> Vector3 a

(*) :: Vector3 a -> Vector3 a -> Vector3 a

negate :: Vector3 a -> Vector3 a

abs :: Vector3 a -> Vector3 a

signum :: Vector3 a -> Vector3 a

fromInteger :: Integer -> Vector3 a

Num a => Num (Vertex3 a)
(Vertex3 a) operator: (-) (+)

Instance details

Methods

(+) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(-) :: Vertex3 a -> Vertex3 a -> Vertex3 a

(*) :: Vertex3 a -> Vertex3 a -> Vertex3 a

negate :: Vertex3 a -> Vertex3 a

abs :: Vertex3 a -> Vertex3 a

signum :: Vertex3 a -> Vertex3 a

fromInteger :: Integer -> Vertex3 a