AronModule

Documentation

class Cnum a where

Overloading the sin, cos to avoid Haskell crazy strict type | TODO: add test cases

Minimal complete definition

_sqrt, _sin, _cos

Methods

_sqrt :: a -> Float

_sin :: a -> Float

_cos :: a -> Float

class Dnum a b where

Division is painful in Haskell

| Try to overload _div for many types | TODO: add more combination? | TODO: add test cases

Minimal complete definition

_div

Associated Types

type DivType a b

Methods

_div :: a -> b -> DivType a b

class Add a b where

Minimal complete definition

addd

Associated Types

type SumTy a b

Methods

addd :: a -> b -> SumTy a b

div' :: Int -> Int -> Double

division is painful like hell in Haskell

mat :: [[Integer]]

simple matrix can be used in GHCi

matr :: [[Ratio Integer]]

mat' :: [[[Char]]]

data Quaternion

Quaternion Implementation

Constructors

Quaternion

Fields

a :: Float

b :: Float

c :: Float

d :: Float

data Complex

Implement complex number operations

Constructors

Complex

Fields

x :: Float

y :: Float

conjugate :: Complex -> Complex

magnitude :: Complex -> Float

realPart :: Complex -> Float

imagPart :: Complex -> Float

coord :: Complex -> (Float, Float)

radianToDegree :: Float -> Float

degreeToRadian :: Float -> Float

tri :: Complex -> (Float, Float, Float)

data GPoint

Constructors

GPoint Int Int

getX :: GPoint -> Int

getY :: GPoint -> Int

toUpperStr :: String -> String

cat :: [String] -> String

replaceList :: [String] -> String -> String -> [String]

["cat", "dog"] "cat" "pig" => ["pig", "dog"]

| replace a matching string with other string

repeat' :: Integer -> a -> [a]

repeat's 3 "pig" => ["pig", "pig", "pig"]

replicate' :: Int -> a -> [a]

replicate' 4 a => "aaaa"

catMaybe :: Maybe [Int] -> Maybe [Int] -> Maybe [Int]

removeSpace :: String -> String

isEmpty :: String -> Bool

trimWS :: String -> String

trim :: String -> String

splitListEmptyLine :: [String] -> [[String]]

split list of string when string is empty/space | ["dog", " ", "", "cat"] => [["dog"], [], ["cat"]]

splitList :: [String] -> ([String], [String])

tuplify2 :: [a] -> (a, a)

tuple to list

tuplify3 :: [a] -> (a, a, a)

ppad :: Show a => [[a]] -> [[[Char]]]

padding whitespace to 2d list

pa :: Show a => [[a]] -> IO ()

pa1 :: Show a => [a] -> IO ()

pm :: Show a => String -> [[a]] -> IO ()

partList :: Int -> [a] -> [[a]]

Fri Dec 7 00:06:35 2018 | partition list to n chunks | Note: print partList n [] will cause error since print needs concrete type | split list to n blocks | partList 2 [1, 2, 3, 4, 5] = [[1,2], [3, 4],[5]]

partList2 :: Int -> [a] -> [[a]]

partitionStrRegex :: String -> String -> [String]

partition string to [String] according to character class [] | r = "[,.]" s = "dog,cat,cow.fox" => ["dog", "cat", "cow", "fox"]

prefix :: String -> [String]

unique :: Ord a => [a] -> [a]

mergeListList :: [[String]] -> [[String]] -> [[String]]

mergeList :: [a] -> [a] -> [a]

mergeListLen :: [a] -> [a] -> Maybe [a]

mergeSortList :: Ord a => [a] -> [a] -> [a]

merge two sorted lists | [1, 4] [3] => [1, 3, 4]

iterateList :: [a] -> (a -> IO ()) -> IO ()

unwrap :: Maybe a -> a

codeCapture :: String -> String

qqsort :: (a -> a -> Bool) -> [a] -> [a]

There are many versions of quicksort here

bad version

quickSort :: [Int] -> [Int]

better version

quickSort' :: [Int] -> [Int]

nice version

quickSort1 :: Ord a => [a] -> [a]

mergeSortC :: (a -> a -> Bool) -> [a] -> [a]

mergeSort :: [Int] -> [Int]

mergeSortM :: (Num a, Eq a) => [[a]] -> [[a]]

readSnippet :: FilePath -> IO [([String], [String])]

read $bvimsnippet.m file | partition blocktexts into list | parse the header. e.g. latex_eqq:*.tex: latex equation | return [([header str]), [blocktext])]

readFileToList :: FilePath -> IO [String]

read dir, all dir or files in [String] | => ["f1", "f2"]

writeToFile :: FilePath -> [String] -> IO ()

old code, if sth break, use the line => readFileToList f = readFile f >>= contents -> return (lines contents) | readFile::FilePath->IO String | fmap line (IO String) => IO String is a Functor | => IO (line String) => IO [String]

wfs :: FilePath -> [String] -> IO ()

wf :: FilePath -> String -> IO ()

writeToFileAppend :: FilePath -> [String] -> IO ()

mToFrac :: (Real a, Fractional b) => [[a]] -> [[b]]

Matrix: Integer to Fractional

class (Num a)=>Fractional a where
class (Num a, Ord a)=> Real where
class (Real a, Enum a)=> Integral where

mToRat :: Real a => [[a]] -> [[Rational]]

Matrix: Real to Rational

data Rational = Ratio Integer

writeToFile2dMat :: (Num a, Fractional a, Show a) => FilePath -> [[a]] -> IO ()

Write 2d Matrix to file, write 2d list to file

\[ \begin{bmatrix} 1 & 2 \\ 3 & 4 \\ \end{bmatrix} \Rightarrow \begin{matrix} 1 & 2 \\ 3 & 4 \\ \end{matrix} \]

m = [[1, 2], [3, 4]]
writeToFile2dMat "/tmp/m.x" m 
:!cat /tmp/m.x

pp :: Show s => s -> IO ()

fpp :: Show s => String -> s -> IO ()

fl :: IO ()

(<<<) :: Show a => String -> a -> String infixl 1

ff :: Show var => String -> var -> IO ()

fw :: String -> IO ()

listDot :: Num a => [a] -> [a] -> a

listDots :: Num a => [[a]] -> [[a]] -> a

scaleList :: Num a => a -> [a] -> [a]

listScale :: Num a => [a] -> a -> [a]

listAdd :: Num a => [a] -> [a] -> [a]

listSub :: Num a => [a] -> [a] -> [a]

listMul :: Num a => [a] -> [a] -> [a]

listNeg :: Num a => [a] -> [a]

dim :: [[a]] -> (Int, Int)

get dimenstion of matrix -> (nrow, ncol) | if [] -> (0, 0)

removeIndex :: Int -> [a] -> [a]

removeRowCol :: Integer -> Integer -> [[a]] -> [[a]]

det :: Num a => [[a]] -> a

Use co-factor expantion to find the determinant of n by n matrix

It is supper slow.

Note: \( \det M = \det M^{T} \) Proof

multiMatInt :: [[Int]] -> [[Int]] -> [[Int]]

matrix multiplicaiton in Int

multiMatDouble :: [[Double]] -> [[Double]] -> [[Double]]

matrix multiplication in Double

multiMatR :: [[Rational]] -> [[Rational]] -> [[Rational]]

matrix multiplication in Rational Number

multiMat :: Num a => [[a]] -> [[a]] -> [[a]]

Use outer product to compute the matrix multiplication

use outer product to compute matrix mulitplication

outer(col_1 row_1) + outer(col_2 row_2) + outer(col_3 row_3)

column vector [[1], [2], [3]] row vector [[1, 2, 3]]

matDiv :: [[Integer]] -> Integer -> [[Integer]]

multiRatMat :: [[String]] -> [[String]] -> [[String]]

matrix multiplication in String/Rational

sumRatList :: [String] -> String

addR :: String -> String -> String

ratToInt :: String -> String

Convert Rational to Integer:

>>> ratToInt "3/1"
3

normalR :: String -> String

normalR -3-1 = 31, 3-2 => -32 |

reduceForm :: [[String]] -> [[String]]

reduceForm list of rational

reduce :: [Integer] -> [Integer]

reduce numerator and denominator:

>>> reduce 6/2
3/1

stringToInt :: String -> Integer

convert string to Integer

>>> stringToInt "123"
123

intToString :: Integer -> String

convert Integer to string

>>> intToString 123
"123"

splitR :: String -> [Integer]

multR :: String -> String -> String

multRL :: String -> [String] -> [String]

normR :: [String] -> [String] -> String

getVec :: Int -> [[a]] -> [[a]]

normList :: Floating a => [[a]] -> [[a]]

Normalize a list as vector

 [[1, 2, 3]]     is row vector
 [[1], [2], [3]] is column vector
 

\[ \| \vec{v} \| = \sqrt{x^2 + y^2 + z^2} \quad \text{ where } \vec{v} = \begin{bmatrix} x & y & z \end{bmatrix} \]

proj :: [String] -> [String] -> [String]

projn :: Fractional a => [[a]] -> [[a]] -> [[a]]

projection from u onto v in n dimensional list | | puv = (u v /v) v | |

rMatrixUpperTri :: Num c => [[[c]]] -> [[[c]]] -> [[c]]

Form a almost R matrix: almost an upper triangle matrix

QR_Decomposition \[ R' = \begin{bmatrix} \left<e_1, v_1 \right> & \left<e_1, v_2 \right> & \left< e_1, v_3 \right> \\ \emptyset & \left<e_2, v_2 \right> & \left< e_2, v_3 \right> \\ \emptyset & \emptyset & \left< e_3, v_3 \right> \end{bmatrix} \quad R = \begin{bmatrix} \left<e_1, v_1 \right> & \left<e_1, v_2 \right> & \left< e_1, v_3 \right> \\ 0 & \left<e_2, v_2 \right> & \left< e_2, v_3 \right> \\ 0 & 0 & \left< e_3, v_3 \right> \end{bmatrix} \\ R' \neq R \]

rejection :: Fractional a => [[[a]]] -> [[[a]]] -> [[[a]]]

rejection \(v_k\) onto span of \( \{ a_1 \dots a_{k-1} \} \)

QR_Decomposition \[ a_k = v_k - \sum_{i=1}^{k-1} \frac{\left<v_k, a_i \right>}{\left<a_i, a_i \right> } a_i \]

v3 projects onto \( \{a_1, a_2 \} \)

a3 = v3 - ((projn v3 a1) + (projn v3 a2)) 

if the matrix is singular then one of \( \left\| a_k \right\| = 0 \)

that can be used to determinate the determinant of a matrix, but not the sign of a determinant

compute the determiant is hard problem: \( \mathcal{O}(n^3) \), there is better algo.

qrDecompose' :: [[Double]] -> ([[Double]], [[Double]])

QR decomposition or QR factorization

QR_Decomposition

Given \(M\) is square matrix, there exists a pair of matrices \( Q \) is unitary matrix and \( R \) is upper triangle matrix such that: \[ M = QR \]

addRL :: [String] -> [String] -> [String]

divR :: String -> String -> String

divR' :: Integer -> Integer -> String

invR :: String -> String

subR :: String -> String -> String

negR :: String -> String

negList :: [String] -> [String]

subRL :: [String] -> [String] -> [String]

diag1 :: Num a => [[a]] -> [a]

get the diagonal of the matrix

uptri :: [[a]] -> [[a]]

get upper triangle of matrix

reverseWord :: String -> String

reverse words in string

rw :: String -> String

tran :: [[a]] -> [[a]]

transpose matrix \[ \begin{bmatrix} a & b \\ c & d \\ \end{bmatrix} \\ \begin{bmatrix} a & c \\ b & d \\ \end{bmatrix} \]

sortRow :: [[Integer]] -> [[Integer]]

concat operator | pp $ "n=" +: fun x | (+:)::(Show a)=> String-> a -> String | (+:) s a = s ++ (show a)

sort row for list of list

upperTri :: [[Integer]] -> [[Integer]]

Tranform the matrix to echolon/upper triangle form

Thu Jan 10 23:29:20 2019

gx http://localhost/pdf/upperTri.pdf

gf UserscatmyfilebitbucketmathupperTri.tex

Note: the determinant is different than the original matrix

Thu Jan 10 23:29:27 2019

Add lcm to the multiplier, make matrix dim from 6 to 13

TODO: fix the num overflow issue

upperTri' :: [[Rational]] -> [[Rational]]

divI :: Fractional a => Integer -> Integer -> a

Division is painful

   class (Num a)=> Fractional a where
       (/):: a -> a -> a

   divI::(Fractional a)=>Integer -> Integer -> a

   class Num a where
       fromInteger::Integer -> a

   class (Num a, Ord a)=> Real a where

   fromIntegral::(Integral a, Num b) => a -> b
   fromIntegral = fromInteger . toInteger

   class (Real a, Enum)=> Integral a where   [1]
   toInteger(Real a, Enum a)=> a -> Integer  [2]
   fromInteger::(Num a)=>Integer -> a        [3]
   
   proof:
   [1] [2] [3] =>
   fromIntegral(Num a) => Integral -> a

   

isInver :: (Fractional a, Ord a) => [[a]] -> Bool

check whether a matrix is singular using QR decomposition

inverse :: [[Double]] -> ([[Double]], [[String]])

Find the invertible matrix, return ([[]], [[]]) if the matrix is singular The code does not check whether the matrix is singular or not

Thu Dec 13 20:38:04 2018 I'm sure there is better way to it TODO: Remove division inverse::(Integral a, Fractional a, Real a, Show a, Eq a)=>[[a]]->([[a]], [[String]]) inverse::(Fractional a, Show a, Eq a)=>[[a]]->([[a]], [[String]])

odds :: [a] -> [a]

evens :: [a] -> [a]

isInvertible :: [[Integer]] -> Bool

------------------------------------------------------------------------------- | Multiply all the diagonal elements and check whether the product is zero or not | --------------------------------------------------------------------------------

ident :: Integer -> [[Integer]]

Generate n dimentional identity matrix

ident' :: Num a => Int -> [[a]]

identS :: Integer -> [[String]]

mlist :: Integer -> [Integer] -> [Integer]

integer multiply integer list | reverse $ 9 x [9, 8] = [8, 8, 2]

drawInt :: Int -> Int -> IO Int

generate Int from x to y random number

drawInteger :: Integer -> Integer -> IO Int

generate Integer from x to y random number

randomDouble :: Int -> IO [Double]

gene list of double from 0 <= x <= 1 | [0.3, 0.4, 0]

randomList :: Int -> IO [Integer]

generate random of list Integer

randomMatrix :: Int -> Int -> IO [[Integer]]

generate m x n random matrix

data Mat

generate zero matrix

Constructors

Zero

Id

geneMat1ToN :: Integer -> [[Integer]]

generate matrix from 1 to n with dimention n | n = 2 | [[1, 2] | [3, 4]] |

geneMat :: Integer -> Mat -> [[Integer]]

generate zero or identity matrix |

help :: IO ()

cmd :: [String] -> IO ()

Add more doc here

compileHaskellToBin :: String -> String -> IO ()

Compile haskell code to $ffmybinmyHaskell => create symbol link $sym/sym | [file:myHaskell.hs] [sym:symbol] link name in $sym

strCompareIC :: String -> String -> Bool

compare two string ignore cases

dirWalk :: FilePath -> (FilePath -> IO [String]) -> IO [String]

walk inside the dir | if there is file, then append it to list | otherwise, walking inside the dir | Pass lambda func: (FilePath -> IO [String]) as argument | | let filefunc fname = | do if (strCompareIC (takeExtension fname) (toUpperStr ".png")) | then return [fname] | else return []

ls :: IO ()

Try to replace as many as shell command as possible | See how far I can go | write many shell commands as possible try to emulate shell commands

lsFile :: String -> IO [String]

list file in the dir = s | NOTE: Not full path

lsFileFull :: String -> IO [String]

lsRegex :: String -> String -> IO [String]

lsFullRegex :: String -> String -> IO [String]

rm :: FilePath -> IO ()

rmDir :: FilePath -> IO ()

pwd :: IO ()

getPwd :: IO FilePath

cd :: FilePath -> IO ()

en :: String -> IO String

cc :: String -> IO ()

g :: IO ()

sleep :: Int -> IO ()

1 sec = 100000

asplitPath :: FilePath -> [String]

len :: (Foldable t, Num b) => t a -> b

better length function

Convert Int to polymorphic values

Convert Int to Num

fromIntegral::(Integral a, Num b)=> a -> b

take' :: Integral n => n -> [a] -> [a]

pathBase :: FilePath -> FilePath

copyRename :: FilePath -> String -> IO ()

copy FilePath to "/tmp" | 2. move the file back to dir with newName | copy file and rename it in the same dir

copyFileToDir :: FilePath -> FilePath -> IO ()

isFile :: FilePath -> IO Bool

fExist :: FilePath -> IO Bool

fe :: FilePath -> IO Bool

createFile :: FilePath -> IO ()

create empty file

copyDir :: FilePath -> FilePath -> IO ()

copy dir to other dir

mv :: FilePath -> FilePath -> IO ()

rename dir or file => dir or file

rename :: FilePath -> FilePath -> IO ()

renameAllFile :: String -> String -> IO ()

mvFiles :: String -> String -> IO ()

mvFile :: FilePath -> FilePath -> IO ()

mkdir :: FilePath -> IO ()

mkdirp :: FilePath -> IO ()

sfilter :: String -> [String] -> [String]

clear :: IO ()

clear terminal screen

run' :: String -> IO ()

run shell cmd, send output to std_out

run :: String -> IO [String]

run shell cmd, capture std_out

df :: Fractional a => (a -> a) -> a -> a

Differentiate on \(f(x)\)

First derivative on f or slop of f at point x

\( f'(x) = \lim_{h \rightarrow 0} \frac{f(x + h) - f(x)}{h} \)

tangent :: Fractional a => (a -> a) -> a -> a -> a

Compute the tangent equation at point \((x_0, f(x_0))\) for function \( f \)

tangent equation:

\( f'(x) = \frac{y - y_0}{x - x_0} \)

\( y = f'(x_0)(x - x_0) + y_0 \)

\( \Rightarrow f'(x_0) = 2x_0 \)

\( f (x) = x^2 \) where \( x = x_0 \)

\( y = 2x_0(x - x_0) + y_0 \) where \( y_0 = f x_0 \)

\( f' (x_0) = \frac{y - y_0)}{x - x0} \)

\( y = f'(x_0)(x - x_0) + y_0 \)

tangentVec :: Fractional a => (a -> a) -> a -> a -> (a, a)

Compute the tangent vector at point (x0, f x0)

prime :: [Integer]

Generate prime number with Sieve Algorithm

oneRoot :: (Double -> Double) -> Double -> Double -> Double -> Maybe Double

------------------------------------------------------------------------------- | Wed Nov 7 18:47:22 2018 | -------------------------------------------------------------------------------- | Example: UserscatmyfilebitbuckethaskellfindRoot.hs | Find root for any function | | Partition the interval [x0, xn] into list = [x0, x1) [x1, x2) [x2, x3) ..[x(n-1), xn) | Cocat [xn, xn] with the list since the left boundary is checked only | Note: f(xn) might be zero, we need to check the xn boundary | TODO: need to check f(xn), DONE, concat [xn, xn] to the list | -------------------------------------------------------------------------------- | Good test cases: | f(x) = x^2 - 4, [-2, 2] | f(x) = x^5 -4*x^4 + 0.1*x^3 + 4*x^2 - 0.5, [-4, 4] → 5 solutions | -------------------------------------------------------------------------------- | limitation: | if each subinterval contains two or more values, then ONLY one value can be found | subinterval can be adjusted in arbitrary small | -------------------------------------------------------------------------------- | 0, 2/2 = [0, 1, 2] = [0, 1) ∪ [1, 2) ∪ [2, 2] | --------------------------------------------------------------------------------

rootList :: (Double -> Double) -> Double -> Double -> Double -> Integer -> [Maybe Double]

Find all the roots for a given close interval: [1, 2], 1 or 2 might be the root |

outer :: Num a => [a] -> [a] -> [[a]]

Sun Oct 21 00:00:05 2018 | outer product = col ⊕ row

outerStr :: (a -> a -> a) -> [[a]] -> [[a]] -> [[a]]

vcol :: [[a]] -> Int -> [[a]]

vrow :: [[a]] -> Int -> [[a]]

zipWith2 :: Num a => (a -> a -> a) -> [[a]] -> [[a]] -> [[a]]

data Tree a

------------------------------------------------------------------------------- | Thu Nov 8 21:24:37 2018 | Most functions are relative Binary Search Tree | --------------------------------------------------------------------------------

Constructors

Empty

Node a (Tree a) (Tree a)

insertNode :: Ord a => Tree a -> a -> Tree a

insertFromList :: Ord a => Tree a -> [a] -> Tree a

inorder :: Tree a -> [a]

maxlen :: Tree a -> Integer

head' :: [a] -> a

TODO: Use better error message | better head

isBST :: Ord a => Tree a -> Bool

------------------------------------------------------------------------------- | check whether a Tree is Binary tree | Thu Nov 8 09:16:34 2018 | -------------------------------------------------------------------------------- | defintion of BST | 1. Null is BST | 2. Left subtree is BST | 3. Right subtree is BST | 4. minimum of left subtree is less than parent node | 5. maximum of right subtree is greater than parent node | --------------------------------------------------------------------------------

binsert :: Tree Integer -> Tree Integer -> Tree Integer

sym :: Tree a -> Bool

lca :: Eq a => Tree a -> a -> a -> Maybe a

------------------------------------------------------------------------------- | Thu Nov 8 16:11:28 2018 | -------------------------------------------------------------------------------- | lease common ancestor | assume two nodes are in the tree | if two nodes are in the same path then the top node will be LCA | --------------------------------------------------------------------------------

buildTree :: [Char] -> [Char] -> Tree Char

Orphan instances

Num a => Num [a]
How to build AronModule.hs | ghc --make AronModule.hs

any dimension list

2*[1, 2] = [2, 4]

[1, 2]+[3, 4] = [4, 6]

2*[[1, 2]] = [[2, 4]]

Instance details

Methods

(+) :: [a] -> [a] -> [a]

(-) :: [a] -> [a] -> [a]

(*) :: [a] -> [a] -> [a]

negate :: [a] -> [a]

abs :: [a] -> [a]

signum :: [a] -> [a]

fromInteger :: Integer -> [a]