AronModule

Documentation

class Cnum a where

Overloading the sin, cos to avoid Haskell crazy strict type | TODO: add test cases

Minimal complete definition

_sqrt, _sin, _cos

Methods

_sqrt :: a -> Float

_sin :: a -> Float

_cos :: a -> Float

class Dnum a b where

Division is painful in Haskell

| Try to overload _div for many types | TODO: add more combination? | TODO: add test cases

Minimal complete definition

_div

Associated Types

type DivType a b

Methods

_div :: a -> b -> DivType a b

class Add a b where

Minimal complete definition

addd

Associated Types

type SumTy a b

Methods

addd :: a -> b -> SumTy a b

div' :: Int -> Int -> Double

division is painful like hell in Haskell

rf :: (Real a, Fractional b) => a -> b

real to Fractional

mat :: [[Integer]]

simple matrix can be used in GHCi

matr :: [[Ratio Integer]]

mat' :: [[[Char]]]

data Quaternion

Quaternion Implementation

Constructors

Quaternion

Fields

a :: Float

b :: Float

c :: Float

d :: Float

eps :: (Num a, Ord a) => a -> a -> a -> Bool

data C

Implement complex number operations

Constructors

C

Fields

x :: Float

y :: Float

con :: C -> C

mag :: C -> Float

re :: C -> Float

im :: C -> Float

co :: C -> (Float, Float)

toPolar :: C -> (Float, Float)

Convert Cartesian Coordinates to Polar Coordinates

toCard :: (Float, Float) -> C

Convert Polar Coordinates to Cartesian Coordinates

sqrtC :: C -> C

square root for Complex number

sqrtC' :: C -> C

radianToDegree :: Float -> Float

degreeToRadian :: Float -> Float

tri :: C -> (Float, Float, Float)

data GPoint

Constructors

GPoint Int Int

getX :: GPoint -> Int

getY :: GPoint -> Int

toUpperStr :: String -> String

concatStr :: [String] -> String -> String

KEY: concat string with delimter string s, concat str, join string

concatStr ["dog", "cat"] [] => "dogcat"
concatStr ["dog", "cat"] " " => "dog cat"
concatStr ["dog", "cat"] ":" => "dog:cat"

cat :: [String] -> String

KEY: cat string, concat string

cat ["dog", "cat"] => "dogcat"

timeNowPico :: IO Integer

(round . (* 10^12)) <$> getPOSIXTime 

timeNowNano :: IO Integer

(round . (* 10^9)) <$> getPOSIXTime 

timeNowMicro :: IO Integer

(round . (* 10^6)) <$> getPOSIXTime 

timeNowMilli :: IO Integer

(round . (* 10^3)) <$> getPOSIXTime 

timeNowSecond :: IO Integer

(round . (* 1)) <$> getPOSIXTime 

getTime :: IO String

KEY: get local current time, local time, time zone

NOTE getCurrentTime is UTC timezone only,

getTimeDay time with day

return $ (show hour) ++ ":" ++ (show minute) ++ ":" ++ (show second)

getLocalDate :: IO String

KEY: get local date, get current time

"2019-05-27 12:57:41.520758 PDT"

containStr :: String -> String -> Bool

Check whether a stirng contains a string pattern

str = "dog cat"
 p = "dog"
containStr str p == return True

replaceList :: [String] -> String -> String -> [String]

replace a matching string with other string

["cat", "dog"] "cat" "pig" => ["pig", "dog"] 

searchReplace :: String -> String -> String -> String

Search replace word, the code is based on matchAllText in package: Text.Regex.TDFA as TD

let s = "mydog dog-- dog pig cat dog fox"
searchReplace s "dog" "[\\0]"
"mydog [dog]-- [dog] pig cat [dog] fox"

Text.Regex => subRegex does not support word boundary

r  = makeRegex ("\\<" ++ word ++ "\\>")::TD.Regex -- word boundary

Regex is from TDFA, NOT Text.Regex >word => "[[:alpha:]]+"

NOTE: support ONE word only, e.g. searchReplace s "dog" "[\0]"

TODO: support multiple words, e.g. searchReplace s "dog|cat" "[\0]"

repeat' :: Integer -> a -> [a]

repeat's 3 "pig" => ["pig", "pig", "pig"]

replicate' :: Int -> a -> [a]

replicate' 4 a => "aaaa"

toStrictBS :: ByteString -> ByteString

Convert Lazy ByteString to Strict ByteString

Data.ByteString => Data.ByteString.Lazy

strictBSToString :: ByteString -> String

Convert Strict ByteString to String

strictBSToString = strictTextToStr . strictByteStringToStrictText

catMaybe :: Maybe [a] -> Maybe [a] -> Maybe [a]

can be replaced with following:

(++) <$> (Just [1]) <*> (Just [2])

removeSpace :: String -> String

isEmpty :: String -> Bool

escapeHtml :: String -> String

Escape Html special characters, e.g. only [<, >] currently

Html special characters can be added ><, >, '\', %

removeIndex :: Int -> [a] -> [a]

Remove element from a list

removeIndex 1 [1,2,3] 
[1, 3]

removeIndex 3 [] 
[]

removeIndex 2 [1]
error

removeIndex 1 "abc"
"ac"

dropEnd :: Integer -> [a] -> [a]

drop n elements from the end

pp $ dropEnd 3 "abcd" == "a"

takeEnd :: Integer -> [a] -> [a]

take n elements from the end

pp $ takeEnd 3 "abcd" == "bcd"

trimWS :: String -> String

trim whitespace from the whole string, NOT just head and tail

see trim head and tail

trimWS s   = filter(not . isSpace) s 

trimAll :: String -> String

trimAll == trimWS, better name:)

trim :: String -> String

trim, remove whitespace characters from either side of string.

see trimWS all whitespace

trimBoth :: String -> String

trimBoth == trim, better name:)

trimStart :: String -> String

Remove whitespace characters from the start of string.

" dog" => "dog"

trimEnd :: String -> String

Remove whitespace characters from the end of string.

" a b c " => " a b c"
"abc"     => "abc"

splitListEmptyLine :: [String] -> [[String]]

split list of string when string is empty/space | ["dog", " ", "", "cat"] => [["dog"], [], ["cat"]]

splitList :: [String] -> ([String], [String])

t1 :: (a, b, c) -> a

t2 :: (a, b, c) -> b

t3 :: (a, b, c) -> c

tuplify2 :: [a] -> (a, a)

tuple to list

tuplify3 :: [a] -> (a, a, a)

isBalanced2 :: String -> String -> (String, Bool)

findBalance :: [(Integer, String)] -> String -> ([(Integer, String)], Bool)

Find all bracket '{, }' until they are balanced

   "{{ {    <- 1
        }   <- 2
       }    <- 3
   }        <- 4
   }"       <- 5
   return  => ([(5,"}")],True)
   

balanceBracket

isBalanced :: String -> Bool

KEY: balance brackets

   pp $ isBalanced "" == True
   pp $ isBalanced "{" == False 
   pp $ isBalanced "}" == False 
   pp $ isBalanced "{}" == True
   pp $ isBalanced "}{" == False 
   pp $ isBalanced "{}{" == False 
   pp $ isBalanced "}{}" == False 
   

balanceBracket

strToLazyText :: String -> Text

String to Lazy Text

lazyTextToStr :: Text -> String

strToStrictText :: String -> Text

KEYS: string to strict text, str to strict text, str to text

strictTextToStr :: Text -> String

Strict Text to String

KEY: strict text to str, text to string

lazyTextToLazyByteString :: Text -> ByteString

lazy Text to lazy ByteString

KEY: lazy Text to lazy ByteString

lazyByteStringToLazyText :: ByteString -> Text

strictTextToStrictByteString :: Text -> ByteString

strictByteStringToStrictText :: ByteString -> Text

Convert Strict ByteString to Strict Text

strictByteStringToLazyByteString :: ByteString -> ByteString

lazyByteStringToStrictByteString :: ByteString -> ByteString

Convert lazy ByteString to strict ByteString

strictTextToLazyText :: Text -> Text

lazyTextToStrictText :: Text -> Text

Lazy Text to Strict Text

strToStrictByteString :: String -> ByteString

String to Strict ByteString, String to ByteString, str to ByteString

strToStrictByteString = strictTextToStrictByteString . strToStrictText

strToLazyByteString :: String -> ByteString

String to lazy ByteString, only for ASCII, not for unicode

   str1 = "好"
   str2 = "good" 

   main = do 
           pp ((LC8.unpack $ LC8.pack str1) == str1) -- False
           pp ((LC8.unpack $ LC8.pack str2) == str2) -- True
   

ppad :: Show a => [[a]] -> [[[Char]]]

padding whitespace to 2d list

pf :: (PrintfArg t1, PrintfType t2) => String -> t1 -> t2

C style printf, string format

import Text.Prinf

 printf "%.2f" 3.0
 3.00
 let n = 3::Int
 let f = rf n
 printf "%.2f, %.2f" f f

let name = "dog"
let weight = 20
printf "(%s)=(%d)"
"(dog)=(20)"

pf "%.2f" 3.0 
3.00

pm :: Show a => String -> [[a]] -> IO ()

Print matrix with labe, print matrix with string,

pm "my matrix" mat
------------------------------------my matrix-----------------------------------
[1,2,3]
[4,5,6]
[7,8,10]

ps :: [String] -> IO ()

print string without double quotes

pa :: Show a => [[a]] -> IO ()

print matrix, show matrix

pre :: (MonadIO m, Show a) => a -> m ()

paa :: Show a => [[a]] -> IO ()

KEY: print matrix, print string without double quotes

partList :: Int -> [a] -> [[a]]

Partition list to n chunks

Note: print partList n [] will cause error since print needs concrete type

split list to n blocks

partList 2 [1, 2, 3, 4, 5] 
[[1,2], [3, 4],[5]]

partList 2 [1, 2, 3, 4, 5, 6] 
[[1,2], [3, 4],[5, 6]]

partList2 :: Int -> [a] -> [[a]]

splitStrChar :: String -> String -> [String]

Partition string to [String] according to character class []

   splitStrChar "[,.]" "dog,cat,cow.fox" => ["dog", "cat", "cow", "fox"]
   splitStrChar::String->String->[String]
   splitStrChar r s = splitWhen(x -> matchTest rex (x:[])) s
               where
                   rex = mkRegex r
   

splitStrRegex => splitStrChar

splitStr :: String -> String -> [String]

KEYS: split string, split str

splitStr "::" "dog::cat" => ["dog", "cat"]

splitBlock :: [String] -> String -> [[String]]

split file with Regex e.g. empty line

   ls <- readFileToList "tmpx.x"
   let lss = splitBlock ls "^[[:space:]]*"  => split file with whitespace
   pp lss

   let lss = splitBlock ls "^[[:space:]]*(---){1,}[[:space:]]*"  => split file "---"
   let lss = splitBlock ls "^[[:space:]]*(===){1,}[[:space:]]*"  => split file "==="
   

TestFile

parseFileBlock :: String -> String -> [String] -> [(Integer, [String])]

parse file and partition it into blocks according to patterns: bs ws

bs = "^[[:space:]]*(---){1,}[[:space:]]*" -- split with delimiter "---"
ws = "[,. ]"                              -- split with [,. ]
parseFileBlock bs ws ["rat", 
                      "-----", 
                      "my dog eats my cat, my cat eats my dog."
                     ]
[(0, ["rat"]), (1, ["my", "dog", "eats", "my", "cat"])]

Test_File

prefix :: String -> [String]

all prefix strings from given string

unique :: Ord a => [a] -> [a]

unique, remove duplicate elements from a list

Convert the list to Set and convert the Set back to List

mergeListList :: [[String]] -> [[String]] -> [[String]]

mergeList :: [a] -> [a] -> [a]

mergeListLen :: [a] -> [a] -> Maybe [a]

mergeSortList :: Ord a => [a] -> [a] -> [a]

merge two sorted lists | [1, 4] [3] => [1, 3, 4]

iterateList :: [a] -> (a -> IO ()) -> IO ()

unwrap :: Maybe a -> a

codeCapture :: String -> String

binarySearch :: Ord a => a -> [a] -> Bool

Binary Search

qqsort :: (a -> a -> Bool) -> [a] -> [a]

quick sort with lambda function

bad version

qqsort(\x y -> len x < len y) ["tiger", "dog"]  => ["dog", "tiger"]

sort :: Ord a => [a] -> [a]

Sort list in nature ordering, e.g. [2, 1] => [1, 2]

sqVec :: Ord a => Vector a -> Vector a

Quick Sort in Vector, Did not see any speed decreasing

quickSort :: [Int] -> [Int]

better version

quickSort' :: [Int] -> [Int]

nice version

quickSort1 :: Ord a => [a] -> [a]

Deprecated

Use quickSortAny

quickSortAny :: Ord a => [a] -> [a]

Quick Sort for any type

same as quickSort1, just a better name

mergeSortC :: (a -> a -> Bool) -> [a] -> [a]

mergeSort :: [Int] -> [Int]

mergeSortM :: (Num a, Eq a) => [[a]] -> [[a]]

groupCount :: [String] -> [(String, Integer)]

groupCount == groupBy . sort

groupCount ["a", "b", "a"] => [("b", 1), ("a", 2)]

See groupBy

readSnippet :: FilePath -> IO [([String], [String])]

Read Snippet

read $b/vim/snippet.m file 
partition blocktexts into list
parse the header: latex_eqq:*.tex: latex equation
return [([latex_eqq:*.tex]), [latex_eqq:*.tex:, latex equation])]

input file:

main:*.hs: my dog, my cat 
my fox eats my rat 

output:
[(["main", "*.hs", "my cat", "my dog"], ["main:*.hs: my dog, my cat ","my fox eats my rat "])]

readFileLatin1ToList :: FilePath -> IO [String]

use latin1 encoding to avoid error when reading non ASCII characters

readFileLatin1 :: FilePath -> IO String

readFileToList :: FilePath -> IO [String]

read dir, all dir or files in [String]

There is error when reading some non utf8 or weird char

Deprecated, use readFileLatin1ToList

 hGetContents: invalid argument (invalid byte sequence)

["f1", "f2"]

writeToFile :: FilePath -> [String] -> IO ()

Write [String] to file

wfs :: FilePath -> [String] -> IO ()

shorthand Write [String] to file

wf :: FilePath -> String -> IO ()

writeToFileAppend :: FilePath -> [String] -> IO ()

nToFractMat :: (Real a, Fractional b) => [[a]] -> [[b]]

Matrix: Integer to Fractional, convert integer matrix to Fractional matrix

see nToNumMat

class (Num a)=>Fractional a where
class (Num a, Ord a)=> Real where
class (Real a, Enum a)=> Integral where

nToNumMat :: Num b => [[Integer]] -> [[b]]

Convert an Integer matrix to an Num Matrix

see nToFractMat

rToRatMat :: Real a => [[a]] -> [[Rational]]

Matrix: Real to Rational

see nToFractMat, nToNumMat

data Rational = Ratio Integer

writeToFile2dMat :: (Num a, Fractional a, Show a) => FilePath -> [[a]] -> IO ()

Write an 2d Matrix to file, write an 2d list to file, write matrix to file

see writeToFileMat

\[ \begin{bmatrix} 1 & 2 \\ 3 & 4 \\ \end{bmatrix} \Rightarrow \begin{matrix} 1 & 2 \\ 3 & 4 \\ \end{matrix} \]

m = [[1, 2], [3, 4]]
writeToFile2dMat "/tmp/m.x" m 
:!cat /tmp/m.x

writeToFileMat :: FilePath -> [[String]] -> IO ()

write String matrix to file

see writeToFile2dMat

filtermap :: (a -> Maybe b) -> [a] -> [b]

filter and map simultaneous

filtermap(\x -> elem x ['a', 'e', 'i', 'o', 'u', 'y'] then Nothing else Just (toUpper x)) "abc"
remove all the vowels and toUpper => "BC"

Prepend an element to a list if available. Leave the list as it is if the first argument is Nothing. Variant of map which deletes elements if the map function returns Nothing.

https://snipplr.com/view/59474/simultaneous-filter-and--map/

putStrLnBS :: ByteString -> IO ()

putStrBS :: ByteString -> IO ()

p :: Show s => s -> IO ()

print output, similar Print.java

pp :: Show s => s -> IO ()

fl :: IO ()

pw :: Show s => String -> s -> IO ()

fpp :: Show s => String -> s -> IO ()

fw :: String -> IO ()

ff :: Show var => String -> var -> IO ()

(<<<) :: Show a => String -> a -> String infixl 1

bs :: [[Double]] -> [Double] -> [Double] -> [Double]

Backward substitute

Given an upper triangle matrix \(A\) and a vector \(\vec{b}\), solve for \(x\)

backward_substitute

\[ Ax = b \] \[ A = \begin{bmatrix} 1 & 2 \\ 0 & 4 \\ \end{bmatrix} \begin{bmatrix} x_1 \\ x_2 \end{bmatrix} = \begin{bmatrix} b_1 \\ b_2 \end{bmatrix} \]

fs :: [[Double]] -> [Double] -> [Double] -> [Double]

forward substitute

Given a lower triangle matrix \(A\) and a vector \(\vec{b}\), solve for \(x\)

backward_substitute

\[ Ax = b \] \[ A = \begin{bmatrix} 1 & 0 \\ 2 & 4 \\ \end{bmatrix} \begin{bmatrix} x_1 \\ x_2 \end{bmatrix} = \begin{bmatrix} b_1 \\ b_2 \end{bmatrix} \]

listDot :: Num a => [a] -> [a] -> a

listDots :: Num a => [[a]] -> [[a]] -> a

scaleList :: Num a => a -> [a] -> [a]

listScale :: Num a => [a] -> a -> [a]

listAdd :: Num a => [a] -> [a] -> [a]

listSub :: Num a => [a] -> [a] -> [a]

listMul :: Num a => [a] -> [a] -> [a]

listNeg :: Num a => [a] -> [a]

dim :: [[a]] -> (Int, Int)

Find the dimension of a matrix -> (nrow, ncol)

dim [] => (0, 0) 

removeRowCol :: Integer -> Integer -> [[a]] -> [[a]]

det :: Num a => [[a]] -> a

Use co-factor expantion to find a determinant of n by n matrix.

It is very slow.

Note: \( \det M = \det M^{T} \) Proof

multiMatInt :: [[Int]] -> [[Int]] -> [[Int]]

matrix multiplicaiton in Int

multiMatDouble :: [[Double]] -> [[Double]] -> [[Double]]

matrix multiplication in Double

multiMatR :: [[Rational]] -> [[Rational]] -> [[Rational]]

matrix multiplication in Rational Number

multiMat :: Num a => [[a]] -> [[a]] -> [[a]]

Use outer product to compute matrix multiplication.

use outer product to compute matrix mulitplication

outer(col_1 row_1) + outer(col_2 row_2) + outer(col_3 row_3)

column vector [[1], [2], [3]] row vector [[1, 2, 3]]

matDiv :: [[Integer]] -> Integer -> [[Integer]]

multiRatMat :: [[String]] -> [[String]] -> [[String]]

matrix multiplication in String/Rational

sumRatList :: [String] -> String

addR :: String -> String -> String

ratToInt :: String -> String

Convert Rational to Integer:

>>> ratToInt "3/1"
3

normalR :: String -> String

normalR -3-1 = 31, 3-2 => -32

reduceForm :: [[String]] -> [[String]]

reduceForm list of rational

reduce :: [Integer] -> [Integer]

reduce numerator and denominator:

>>> reduce 6/2
3/1

stringToInt :: String -> Integer

Convert string to Integer, str to int, str to num, string to num

stringToInt "123"
123

stringToInt "a"
error

word :: String -> (String, String)

split first word by delimiter: isSpace

split "my dog eats my cat"
("my", "dog eats my cat")

parser

isWord :: String -> Bool

Check a string whether it is a word or not

 Word is defined as  "^[[:alpha:]]+$"
isWord "dog"
True

isWord "dog1"
False

isWord::String -> Bool
isWord s = matchTest (mkRegex "^[[:alpha:]]+$") s 

sentence :: String -> [String]

split all words with delimiter: isSpace

parser

print $ sentence "dog cat 133"
["dog", "cat", "133"]

perm :: [a] -> [[[a]]]

Permutation of any type, the algo is based on prefix string

permutation_tree

   fun (prefix, int index, removeIndex(index str))
   

 perm "12"
[["12", "21"]]

perm2 :: Eq a => [a] -> [[a]]

Permutation of any type

 perm "12"
[["12", "21"]]

       abc
           a   bc
               b c
                   [[]]
               c b
                   [[]]

           b   ac
               a c
                   [[]]
               c a
                   [[]]

           c   ab
               a b
                   [[]]
               b a
                   [[]]
   

perm3 :: [a] -> [[a]]

And other Permutation

SO

intToString :: Integer -> String

Convert Integer to string, int to str

>>> intToString 123
"123"

splitR :: String -> [Integer]

multR :: String -> String -> String

multRL :: String -> [String] -> [String]

normR :: [String] -> [String] -> String

getVec :: Int -> [[a]] -> [[a]]

normList :: Floating a => [[a]] -> [[a]]

Normalize a list as vector

 [[1, 2, 3]]     is row vector
 [[1], [2], [3]] is column vector
 

\[ \| \vec{v} \| = \sqrt{x^2 + y^2 + z^2} \quad \text{ where } \vec{v} = \begin{bmatrix} x & y & z \end{bmatrix} \]

proj :: [String] -> [String] -> [String]

projn :: Fractional a => [[a]] -> [[a]] -> [[a]]

projection from u onto v in n dimensional list | | puv = (u v /v) v | |

rMatrixUpperTri :: Num c => [[[c]]] -> [[[c]]] -> [[c]]

Form a almost R matrix: almost an upper triangle matrix

QR_Decomposition \[ R' = \begin{bmatrix} \left<e_1, v_1 \right> & \left<e_1, v_2 \right> & \left< e_1, v_3 \right> \\ \emptyset & \left<e_2, v_2 \right> & \left< e_2, v_3 \right> \\ \emptyset & \emptyset & \left< e_3, v_3 \right> \end{bmatrix} \quad R = \begin{bmatrix} \left<e_1, v_1 \right> & \left<e_1, v_2 \right> & \left< e_1, v_3 \right> \\ 0 & \left<e_2, v_2 \right> & \left< e_2, v_3 \right> \\ 0 & 0 & \left< e_3, v_3 \right> \end{bmatrix} \\ R' \neq R \]

rejection :: Fractional a => [[[a]]] -> [[[a]]] -> [[[a]]]

rejection \(v_k\) onto span of \( \{ a_1 \dots a_{k-1} \} \)

QR_Decomposition \[ a_k = v_k - \sum_{i=1}^{k-1} \frac{\left<v_k, a_i \right>}{\left<a_i, a_i \right> } a_i \]

v3 projects onto \( \{a_1, a_2 \} \)

a3 = v3 - ((projn v3 a1) + (projn v3 a2)) 

if the matrix is singular then one of \( \left\| a_k \right\| = 0 \)

that can be used to determinate the determinant of a matrix, but not the sign of a determinant

compute the determiant is hard problem: \( \mathcal{O}(n^3) \), there is better algo.

qrDecompose' :: [[Double]] -> ([[Double]], [[Double]])

QR decomposition or QR factorization

QR_Decomposition

Given \(M\) is square matrix, there exists a pair of matrices \( Q \) is unitary matrix and \( R \) is upper triangle matrix such that: \[ M = QR \]

addRL :: [String] -> [String] -> [String]

Addition for two list of Rational strings

addRL ["1/2", "1/2"] ["1/2", "1/2"]
["1/1", "1/1"]

divR :: String -> String -> String

Division for Rational string

divR "1/2" "3/4"
"2/3"

divR' :: Integer -> Integer -> String

invR :: String -> String

subR :: String -> String -> String

negR :: String -> String

negList :: [String] -> [String]

subRL :: [String] -> [String] -> [String]

eightQueen :: [[Integer]] -> [(Integer, Integer)] -> Integer -> Integer -> [[Integer]]

The algorithm uses backtrack.

Move top left corner to right and down

IF the move is valid THEN set (c, r) = 1, AND move one level down to (c+1, 0)

IF the move is not valid THEN move to (c, r+1)

Move (c,r) from left to right

IF (c,r) is valid move

THEN go down one level deep, => (c+1,0)

ELSE goto (c, r+1)

IF (c, r) is the right most position, AND IF (c,r) is valid move

THEN go down to (c+1, 0)

ELSE backtrack => take the previous valid move and reset (c,r)=0 AND check whether (c, r+1) is valid move.

IF (c, 0) is the bottom level,

THEN we are done!

let m1 = geneMat 10 Zero
pa $ eightQueen m1 [] 0 0 

sumLeft :: [[Integer]] -> Integer -> Integer -> Integer

sumRight :: [[Integer]] -> Integer -> Integer -> Integer

rep2d :: (Integral n1, Integral n2) => [[a]] -> n1 -> n2 -> a -> [[a]]

Replace element in position (c, r) with element n in 2d matrix. replace element, replace matrix, replace 2d array

diag1 :: Num a => [[a]] -> [a]

get the diagonal of the matrix

uptri :: [[a]] -> [[a]]

get upper triangle of matrix

leftDiagonal :: [[Integer]] -> Integer -> Integer -> [Integer]

Collect all the elements from left diagonal of a matrix.

   mat = [
         [1, 2, 3],
         [4, 5, 6],
         [7, 8, 10]
         ]
   leftDiagonal mat 0 0
   [7, 5, 3]
   

rightDiagonal :: [[Integer]] -> Integer -> Integer -> [Integer]

Collect all the elements from right diagonal of a matrix.

   mat = [
         [1, 2, 3],
         [4, 5, 6],
         [7, 8, 10]
         ]
   rightDiagonal mat 0 0
   [10, 5, 1]
   

reverseWord :: String -> String

reverse words in string

rw :: String -> String

tran :: [[a]] -> [[a]]

Transpose matrix

\[ \begin{bmatrix} a & b \\ c & d \\ \end{bmatrix} \Rightarrow \begin{bmatrix} a & c \\ b & d \\ \end{bmatrix} \]

sortRow :: [[Integer]] -> [[Integer]]

concat operator | pp $ "n=" +: fun x | (+:)::(Show a)=> String-> a -> String | (+:) s a = s ++ (show a)

sort row for list of list

upperTri :: [[Integer]] -> [[Integer]]

Tranform the matrix to echolon/upper triangle form

Thu Jan 10 23:29:20 2019

gx upperTri

gf UserscatmyfilebitbucketmathupperTri.tex

Note: the determinant is different than the original matrix

Thu Jan 10 23:29:27 2019

Add lcm to the multiplier, make matrix dim from 6 to 13

TODO: fix the num overflow issue

upperTri' :: [[Rational]] -> [[Rational]]

divI :: Fractional a => Integer -> Integer -> a

Division is painful

   class (Num a)=> Fractional a where
       (/):: a -> a -> a

   divI::(Fractional a)=>Integer -> Integer -> a

   class Num a where
       fromInteger::Integer -> a

   class (Num a, Ord a)=> Real a where

   fromIntegral::(Integral a, Num b) => a -> b
   fromIntegral = fromInteger . toInteger

   class (Real a, Enum)=> Integral a where   [1]
   toInteger(Real a, Enum a)=> a -> Integer  [2]
   fromInteger::(Num a)=>Integer -> a        [3]
   
   proof:
   [1] [2] [3] =>
   fromIntegral(Num a) => Integral -> a

   

inverse :: [[Double]] -> ([[Double]], [[String]])

Find the invertible matrix, return ([[]], [[]]) if the matrix is singular

The code does not check whether the matrix is singular or not

Thu Dec 13 20:38:04 2018 I'm sure there is better way to it

m <- randomMatrix 7 7  -- Int matrix
m' = (map . map) rf m  -- Double matrix
inverse m'

inverse::(Integral a, Fractional a, Real a, Show a, Eq a)=>[[a]]->([[a]], [[String]])
inverse::(Fractional a, Show a, Eq a)=>[[a]]->([[a]], [[String]])

TODO: Remove division

odds :: [a] -> [a]

evens :: [a] -> [a]

isInver :: (Fractional a, Ord a) => [[a]] -> Bool

check whether a matrix is singular using QR_Decompoisition

isInvertible :: [[Integer]] -> Bool

Multiply all the diagonal elements and check whether the product is zero or not

Algorithm:

Find the upper triangle of m

Multiply all the diagonal entries

If their product is NOT zero then it is invertible, otherwise singular

   -- Test case:
   rm <- randomMatrix 100 100
   pp $ isInvertible rm
   

ident :: Integer -> [[Integer]]

Generate n dimentional identity matrix

ident' :: Num a => Int -> [[a]]

identS :: Integer -> [[String]]

mlist :: Integer -> [Integer] -> [Integer]

integer multiply integer list | reverse $ 9 x [9, 8] = [8, 8, 2]

drawInt :: Int -> Int -> IO Int

generate Int from x to y random number

drawInteger :: Integer -> Integer -> IO Int

generate Integer from x to y random number

randomDouble :: Int -> IO [Double]

Generate list of Double in \( x \in [0 \dots n], n = 100 \)

[0.3, 0.4, 0]

Use randomFrac

randomFloat :: Int -> IO [Float]

Generate list of Double in \( x \in [0 \dots n], n = 100 \)

[0.3, 0.4, 0]

Use randomFrac

randomFrac :: Fractional a => Int -> IO [a]

Generate list of Fractional in \( x \in [0 \dots n], n = 100 \)

[0.3, 0.4, 0]

Num and Fractional are type class so you can NOT do [Num] or [Fractional]

Type class in Haskell is similar like interface in Java, but there are many differences

e.g. Integer implement Num => Integer is a concrete class e.g. Fractional inherits from Num => Fractional type class e.g. Float implement Fractional => Float is concrete class

randomList :: Int -> IO [Integer]

generate random of list Integer

randomMatrix :: Num a => Int -> Int -> IO [[a]]

Generate \( m \times n \) random matrix.

data Mat

generate zero matrix

Constructors

Zero

Id

geneMat1ToN :: Integer -> [[Integer]]

Generate a matrix from 1 to \( n \times n \) with dimention n

geneMat1ToN 2
[[1, 2],
 [3, 4]]

geneMatMN :: Integer -> Integer -> [[Integer]]

Generate a matrix from 1 to \( m \times n \)

geneMat1ToN 2 3 
[[1, 2, 3],
 [4, 5, 6]]

geneMat :: Integer -> Mat -> [[Integer]]

generate zero or identity matrix |

help :: IO ()

cmd :: [String] -> IO ()

Add more doc here

compileHaskellToBin :: String -> String -> IO ()

Compile haskell code to $ffmybinmyHaskell => create symbol link $sym/sym | [file:myHaskell.hs] [sym:symbol] link name in $sym

strCompareIC :: String -> String -> Bool

compare two string ignore cases

baseName :: FilePath -> String

file base name

KEY: file extension, extension, basename, base name, file ext

baseName "/dog/file.txt" => "file"

takeFileName gives "file.ext"
takeDirectory gives "/directory"
takeExtension gives ".ext"
dropExtension gives "/directory/file"
takeBaseName gives "file"
"/directory" </> "file.ext".
"/directory/file" <.> "ext".
"/directory/file.txt" -<.> "ext".

gotoCurrDir :: IO ()

Goto current directory, use in script

dirWalk :: FilePath -> (FilePath -> IO [String]) -> IO [String]

copy directory with lambda function: upper case file extension

if there is file, then append it to list

otherwise, walking inside the dir

Pass lambda func: (FilePath -> IO [String]) as argument

let filefunc fname = 
     do if (strCompareIC (takeExtension fname) (toUpperStr ".png"))
          then return [fname]
          else return []

ls :: IO ()

Try to replace as many as shell command as possible | See how far I can go | write many shell commands as possible try to emulate shell commands

lsFile :: String -> IO [String]

list file in the dir = s | NOTE: Not full path

lsFileFull :: String -> IO [String]

lsRegex :: String -> String -> IO [String]

list file with regex match, see lsRegexFull, list file with filter, file filter

ff <- lsRegex (getEnv j) "\\.java$" -- list all java file
ff <- lsRegex (getEnv j) "^try"     -- list all file names starts with "try"

lsRegexFull :: String -> String -> IO [String]

list full path with regex match, see lsRegex, list file with filter, file filter, file path with regex

lsRegexFull "." "\\.hs$"
lsRegexFull "/tmp" "\\.hs$"

lsFullRegex :: String -> String -> IO [String]

Deprecated, use lsRegexFull

list full path with regex match

lsRegexFull "." "\\.hs"
lsRegexFull "/tmp" "\\.hs"

rm :: FilePath -> IO ()

remove file only

rmDir :: FilePath -> IO ()

remove directory recursive

rm "dir"

pwd :: IO ()

getPwd :: IO FilePath

get current dir

cd :: FilePath -> IO ()

change dir

en :: String -> IO String

getEnt

cc :: String -> IO ()

g :: IO ()

sleep :: Int -> IO ()

sleep 1 sec = 100000

asplitPath :: FilePath -> [String]

split path

asplitPath "/dot/cat/"
["dog", "cat"]

asplitPath "dot/cat/"
["dog", "cat"]

asplitPath "/dot/cat"
["dog", "cat"]

asplitPath "dot/cat"
["dog", "cat"]

splitPathA :: FilePath -> [FilePath]

splitPathA from System.FilePath.Posix.splitPath

splitPathA "/dog/cat/"
["/", "dog/", "cat/"]

splitPathA "/dog/cat"
["/", "dog/", "cat"]

splitPathA "dog/cat"
["dog/", "cat"]

splitPathA "/"
["/"]

splitPathA "./"
["./"]

len :: (Foldable t, Num b) => t a -> b

better length function

Convert Int to polymorphic values

Convert Int to Num

fromIntegral::(Integral a, Num b)=> a -> b

take' :: Integral n => n -> [a] -> [a]

drop' :: Integral n => n -> [a] -> [a]

pathBase :: FilePath -> FilePath

copyRename :: FilePath -> String -> IO ()

copy FilePath to "/tmp" | 2. move the file back to dir with newName | copy file and rename it in the same dir

copyFileToDir :: FilePath -> FilePath -> IO ()

isFile :: FilePath -> IO Bool

fExist :: FilePath -> IO Bool

fe :: FilePath -> IO Bool

createFile :: FilePath -> IO ()

create empty file

listDirFilter :: FilePath -> String -> IO [FilePath]

list file with filter

e.g. all html files >ls <- listDirFilter pa "\.html$"

copyDir :: FilePath -> FilePath -> IO ()

copy dir to other dir

mv :: FilePath -> FilePath -> IO ()

rename dir or file => dir or file

rename :: FilePath -> FilePath -> IO ()

renameAllFile :: String -> String -> IO ()

mvFiles :: String -> String -> IO ()

mvFile :: FilePath -> FilePath -> IO ()

mkdir :: FilePath -> IO ()

mkdirp :: FilePath -> IO ()

sfilter :: String -> [String] -> [String]

clear :: IO ()

clear terminal screen

run' :: String -> IO ()

run shell cmd, send output to std_out

sys :: String -> IO ExitCode

Do not need to return IO [String] like run

or use :! cmd in GHCi

run :: String -> IO [String]

Run shell cmd, capture std_out

some issue with waitForProcess

it might be deadlock, e.g. run "ps aux"

man ps -x => still d't underand why the process is not terminated

top command issue => process is not terminated

Change it to System, there is issue with the function TODO: fix it:o

df :: Fractional a => (a -> a) -> a -> a

Differentiate on \(f(x)\)

Find the first derivative at \( x \) on function \( (a \rightarrow a)) \)

x = 1.0
df (\x ->x^2) x
2.0

\( f(x) = x^2 \Rightarrow f'(x) = 2x \Rightarrow f'(1.0) = 2.0 \)

First derivative on f or slop of f at point x

\( f'(x) = \lim_{h \rightarrow 0} \frac{f(x + h) - f(x)}{h} \)

tangent :: Fractional a => (a -> a) -> a -> a -> a

Compute the tangent equation at point \((x_0, f(x_0))\) for function \( f \)

Find tangent equation at \( x_0 = 1.0 \) in \( f(x) = x^2 \)

tanglent (\x -> x^2) 0 1

tangent equation:

\( f'(x) = \frac{y - y_0}{x - x_0} \)

\( y = f'(x_0)(x - x_0) + y_0 \)

e.g.

\( f(x) = x^2 \)

\( \Rightarrow f'(x_0) = 2x_0 \)

\( f (x) = x^2 \) where \( x = x_0 \)

\( y = 2x_0(x - x_0) + y_0 \) where \( y_0 = f (x_0) \)

\( f' (x_0) = \frac{y - y_0}{x - x0} \)

\( y = f'(x_0)(x - x_0) + y_0 \)

tangentVec :: Fractional a => (a -> a) -> a -> a -> (a, a)

Compute the tangent vector at point (x0, f x0)

prime :: [Integer]

Generate prime number with Sieve Algorithm

oneRoot :: (Double -> Double) -> Double -> Double -> Double -> Maybe Double

Find root for any polynomial function

Example: FindRoot

Partition the interval [x0, xn] into list = [x0, x1) [x1, x2) [x2, x3) ..[x(n-1), xn)

Concat [xn, xn] with the list since the left boundary is checked only

Note: f(xn) might be zero, we need to check the xn boundary

TODO: need to check f(xn), DONE, concat [xn, xn] to the list

Good test cases:

\( f(x) = x^2 - 4 \quad x \in [-2, 2] \)

\( f(x) = x^5 -4x^4 + 0.1x^3 + 4x^2 - 0.5 \quad x \in [-4, 4] \Rightarrow \) 5 solutions

limitation:

if each subinterval contains two or more values, then ONLY one value can be found

subinterval can be adjusted in arbitrary small

[0, 2] (2-0)/2 = [0, 1, 2] = [0, 1) ∪ [1, 2) ∪ [2, 2]

rootList :: (Double -> Double) -> Double -> Double -> Double -> Integer -> [Maybe Double]

Find all the roots for a given close interval: [1, 2], 1 or 2 might be the root |

outer :: Num a => [a] -> [a] -> [[a]]

Sun Oct 21 00:00:05 2018 | outer product = col ⊕ row

outerStr :: (a -> a -> a) -> [[a]] -> [[a]] -> [[a]]

vcol :: [[a]] -> Int -> [[a]]

vrow :: [[a]] -> Int -> [[a]]

zipWith2 :: Num a => (a -> a -> a) -> [[a]] -> [[a]] -> [[a]]

zipWith in two dimensions, zipWith matrix

   zipWith2(Num a) => (a -> a -> a) ->[[a]] ->[[a]] -> [[a]]
   zipWith2 f a b = [ zipWith f ra rb | (ra, rb) <- zip a b]

   zipWith2::(Num a)=>(a ->a ->a)->[[a]]->[[a]]->[[a]]
   zipWith2 f m1 m2 = zipWith(x y -> zipWith(x1 y1 -> f x1 y1) x y) m1 m2 
   

data XNode

Constructors

XNode (HashMap Char XNode) Bool

insertTries :: String -> XNode -> XNode

insert operation for Tries data structure

let xn = insertTries "a" (XNode M.empty False)
let xxn = insertTries"ab" xn
pp $ "containsTries=" <<< (containsTries "a" xxn == True)
pp $ "containsTries=" <<< (containsTries "ab" xxn == True)

If String is empty return XNode is the end of a word

If x points to Nothing(x is not on the Tries), then recur to cx

If x points to an XNode, then recur to cx

insertTriesList :: [String] -> XNode -> XNode

Insert list of strings to Tries, see insertTries

containsTries :: String -> XNode -> Bool

contain operation for Tries data structure

let xn = insertTries "a" (XNode M.empty False)
let xxn = insertTries"ab" xn
pp $ "containsTries=" <<< (containsTries "a" xxn == True)
pp $ "containsTries=" <<< (containsTries "ab" xxn == True)

data Tree a

------------------------------------------------------------------------------- | Thu Nov 8 21:24:37 2018 | Most functions are related to Binary Search Tree | --------------------------------------------------------------------------------

Constructors

Empty

Node a (Tree a) (Tree a)

insertNode :: Ord a => Tree a -> a -> Tree a

insertFromList :: Ord a => Tree a -> [a] -> Tree a

inorder :: Tree a -> [a]

maxlen :: Tree a -> Integer

head' :: [a] -> a

TODO: Use better error message | better head

isBST :: Ord a => Tree a -> Bool

check whether a Tree is Binary tree

defintion of BST

Null is BST

Left subtree is BST

Right subtree is BST

minimum of left subtree is less than parent node

maximum of right subtree is greater than parent node

binsert :: Tree Integer -> Tree Integer -> Tree Integer

Binary tree insection

sym :: Tree a -> Bool

lca :: Eq a => Tree a -> a -> a -> Maybe a

Lease common ancestor

assume two nodes are in the tree

if two nodes are in the same path then the top node will be LCA

buildTree :: [Char] -> [Char] -> Tree Char

Build binary tree from preorder and inorder

anagram :: String -> [String] -> [String]

Find all anagrams from a list of strings

anagram "dog" ["god", "cat", "ogd"] 
["god", "ogd"]

redisExtractAronModule :: [String] -> [([String], Integer, [String])]

redisExtractJavaMethod :: [String] -> [([String], Integer, [String])]

KEY: Parse Java method name, extract method name and form a list [([String], Integer, [String])]

File format: jname = "UserscatmyfilebitbucketjavalibAron.java"

The list can be used in Redis Server

["line1", "line2"] -> [([k0, k1], 1, ["line1"])]

textArea :: Integer -> Integer -> String -> String

Html textarea textArea row col string textArea 4 5 "dog"

Orphan instances

Num a => Num [a]
How to build AronModule.hs | ghc --make AronModule.hs

any dimension list

2*[1, 2] = [2, 4]

[1, 2]+[3, 4] = [4, 6]

2*[[1, 2]] = [[2, 4]]

Instance details

Methods

(+) :: [a] -> [a] -> [a]

(-) :: [a] -> [a] -> [a]

(*) :: [a] -> [a] -> [a]

negate :: [a] -> [a]

abs :: [a] -> [a]

signum :: [a] -> [a]

fromInteger :: Integer -> [a]